### 2025-03-12 #7894 @editorialbot
[PRE REVIEW]: modelbased: An R package to make the most out of your statistical models through marginal means, marginal effects, and model predictions
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** https://github.com/easystats/modelbased
**Branch with paper.md** (empty if default branch):
**Version:** v0.10.0
