### 2025-03-12 #7894 @arfon
@DominiqueMakowski – thanks for your submission to JOSS. While we work to find an editor, please work to reduce the length of your paper. JOSS papers are expected to be ~1000 words.
