### 2025-03-12 #7894 @DominiqueMakowski
Thanks @arfon !

Potential reviewers:

- vincentarelbundock
- jmgirard
- tjmahr
- profandyfield
- bbolker
- singmann
