+++ R/estimate_grouplevel.R
@@ -54,7 +54,18 @@
 #' # overall coefficients
 #' estimate_grouplevel(model, type = "total")
 #' @export
-estimate_grouplevel <- function(model, type = "random", dispersion = TRUE, test = NULL, diagnostic = NULL, ...) {
+estimate_grouplevel <- function(model, ...) {
+  UseMethod("estimate_grouplevel")
+}
+
+
+#' @export
+estimate_grouplevel.default <- function(model,
+                                        type = "random",
+                                        dispersion = TRUE,
+                                        test = NULL,
+                                        diagnostic = NULL,
+                                        ...) {
   # validate argument
   type <- insight::validate_argument(type, c("random", "total"))
 
@@ -102,42 +113,154 @@ estimate_grouplevel <- function(model, type = "random", dispersion = TRUE, test 
   # Remove columns with only NaNs (as these are probably those of fixed effects)
   random[vapply(random, function(x) all(is.na(x)), TRUE)] <- NULL
 
-  # Clean
-  row.names(random) <- NULL
-  random$Effects <- NULL
-  if ("Component" %in% names(random) && insight::n_unique(random$Component) == 1 && unique(random$Component) == "conditional") {
-    random$Component <- NULL
+  # Clean and Reorganize columns
+  random <- .clean_grouplevel(random)
+
+  # Sort
+  random <- .sort_random_effects(random)
+
+  # add data and attributes
+  .add_grouplevel_attributes(model, type, params, random)
+}
+
+
+#' @export
+estimate_grouplevel.brmsfit <- function(model,
+                                        type = "random",
+                                        dispersion = TRUE,
+                                        test = NULL,
+                                        diagnostic = NULL,
+                                        ...) {
+  # validate argument
+  type <- insight::validate_argument(type, c("random", "total"))
+
+  # sanity check
+  if (is.null(insight::find_random(model))) {
+    insight::format_error("Model must be a mixed model with random effects.")
   }
 
-  # Reorganize columns
-  random <- datawizard::data_relocate(random, c("Component", "Group", "Level", "Parameter"), verbose = FALSE)
+  # Extract params
+  params <- parameters::model_parameters(
+    model,
+    effects = ifelse(type == "random", "all", "total"),
+    group_level = identical(type, "random"),
+    dispersion = dispersion,
+    test = test,
+    diagnostic = diagnostic,
+    ...
+  )
+
+  # get cleaned parameter names with additional information
+  clean_parameters <- attributes(params)$clean_parameters
+
+  # Re-add info
+  if (!"Group" %in% names(params) && !is.null(clean_parameters)) {
+    params$Group <- clean_parameters$Group
+  }
+  if (!"Level" %in% names(params) && !is.null(clean_parameters)) {
+    params$Level <- clean_parameters$Cleaned_Parameter
+  }
+
+  # TODO: improve / add new printing that groups by group/level?
+  random <- as.data.frame(params[params$Effects == type, ])
+
+  # Remove columns with only NaNs (as these are probably those of fixed effects)
+  random[vapply(random, function(x) all(is.na(x)), TRUE)] <- NULL
+
+  # Clean and Reorganize columns
+  random <- .clean_grouplevel(random)
 
   # Clean-up brms output
-  if (inherits(model, "brmsfit") && type == "random") {
+  if (type == "random") {
     # Save brms name (just in case)
     random$Name <- random$Parameter
     # Filter out non-random effects
     random <- random[startsWith(random$Parameter, "r_"), ]
     # Remove Group from Level
[…]
