+++ man/estimate_grouplevel.Rd
@@ -5,19 +5,21 @@
 \alias{reshape_grouplevel}
 \title{Group-specific parameters of mixed models random effects}
 \usage{
-estimate_grouplevel(
-  model,
-  type = "random",
-  dispersion = TRUE,
-  test = NULL,
-  diagnostic = NULL,
-  ...
-)
+estimate_grouplevel(model, ...)
 
 reshape_grouplevel(x, indices = "all", group = "all", ...)
 }
 \arguments{
 \item{model}{A mixed model with random effects.}
+
+\item{...}{Other arguments passed to \code{\link[parameters:model_parameters]{parameters::model_parameters()}}.}
+
+\item{x}{The output of \code{estimate_grouplevel()}.}
+
+\item{indices}{A list containing the indices (i.e., which columns) to
+extract (e.g., "Coefficient").}
+
+\item{group}{A list containing the random factors to select.}
 
 \item{type}{\code{"random"} or \code{"total"}. If \code{"random"} (default), the coefficients correspond to the
 conditional estimates of  the random effects (as they are returned by
@@ -36,15 +38,6 @@ possible to compute them.}
 \item{dispersion, test, diagnostic}{Arguments passed to \code{\link[parameters:model_parameters]{parameters::model_parameters()}} for
 Bayesian models. By default, it won't return significance or diagnostic indices
 (as it is not typically very useful).}
-
-\item{...}{Other arguments passed to \code{\link[parameters:model_parameters]{parameters::model_parameters()}}.}
-
-\item{x}{The output of \code{estimate_grouplevel()}.}
-
-\item{indices}{A list containing the indices (i.e., which columns) to
-extract (e.g., "Coefficient").}
-
-\item{group}{A list containing the random factors to select.}
 }
 \description{
 Extract random parameters of each individual group in the context of mixed
