+++ NAMESPACE
@@ -5,6 +5,9 @@ S3method(describe_nonlinear,estimate_predicted)
 S3method(describe_nonlinear,numeric)
 S3method(estimate_contrasts,default)
 S3method(estimate_contrasts,estimate_predicted)
+S3method(estimate_grouplevel,brmsfit)
+S3method(estimate_grouplevel,default)
+S3method(estimate_grouplevel,stanreg)
 S3method(format,estimate_contrasts)
 S3method(format,estimate_grouplevel)
 S3method(format,estimate_means)
