+++ R/estimate_grouplevel.R
@@ -224,15 +224,16 @@ estimate_grouplevel.stanreg <- function(model,
   ## TODO: for now, rstanarm has no random effects on the sigma parameter
   ## however, if this changes, we need another solution here (and in
   ## insight::clean_parameter())
+  if (!is.null(clean_parameters)) {
+    # fix for rstanarm, which contains a sigma columns
+    clean_parameters <- clean_parameters[
+      clean_parameters$Component != "sigma" & !startsWith(clean_parameters$Parameter, "Sigma["), # nolint
+    ]
 
-  # fix for rstanarm, which contains a sigma columns
-  clean_parameters <- clean_parameters[
-    clean_parameters$Component != "sigma" & !startsWith(clean_parameters$Parameter, "Sigma["), # nolint
-  ]
-
-  params$Parameter <- insight::trim_ws(sub(":.*", "", clean_parameters$Group))
-  params$Group <- insight::trim_ws(sub("^[^:]*:", "", clean_parameters$Group))
-  params$Level <- insight::trim_ws(sub("^[^:]*:", "", clean_parameters$Cleaned_Parameter))
+    params$Parameter <- insight::trim_ws(sub(":.*", "", clean_parameters$Group))
+    params$Group <- insight::trim_ws(sub("^[^:]*:", "", clean_parameters$Group))
+    params$Level <- insight::trim_ws(sub("^[^:]*:", "", clean_parameters$Cleaned_Parameter))
+  }
 
   # TODO: improve / add new printing that groups by group/level?
   random <- as.data.frame(params[params$Effects == type, ])
