### 2025-03-12 2582483
estimate_grouplevle(): return dispersion for Bayesian models (#443)
* Dispersion = TRUE by default + allow more flexibility in cols returned

* docs

* `estimate_grouplevel()` bug (#446)

* add comment

* update news

* Update NEWS.md

Co-authored-by: Dominique Makowski <dom.mak19@gmail.com>

* fix

* fix, add test

* update tests

* minor cleaning

* methods

* clean

* wordlist

* finalize

* fix

* fix

---------

Co-authored-by: Daniel <mail@danielluedecke.de>
