+++ R/estimate_grouplevel.R
@@ -6,20 +6,24 @@
 #' which can be useful to add the random effects to the original data.
 #'
 #' @param model A mixed model with random effects.
-#' @param type `"random"` or `"total"`. If `"random"` (default), the coefficients correspond to the
-#'   conditional estimates of  the random effects (as they are returned by
-#'   `lme4::ranef()`). They typically correspond to the deviation of each
-#'   individual group from their fixed effect (assuming the random effect is
-#'   also included as a fixed effect). As such, a coefficient close to 0
-#'   means that the participants' effect is the same as the population-level
-#'   effect (in other words, it is "in the norm"). If `"total"`, it will return
-#'   the sum of the random effect and its corresponding fixed effects, which
-#'   internally relies on the `coef()` method (see `?coef.merMod`). Note that
-#'   `type = "total"` yet does not return uncertainty indices (such as SE and CI)
-#'   for models from *lme4* or *glmmTMB*, as the necessary information to
-#'   compute them is not yet available. However, for Bayesian models, it is
-#'   possible to compute them.
-#' @param ... Other arguments passed to or from other methods.
+#' @param type `"random"` or `"total"`. If `"random"` (default), the
+#'   coefficients correspond to the conditional estimates of  the random effects
+#'   (as they are returned by `lme4::ranef()`). They typically correspond to the
+#'   deviation of each individual group from their fixed effect (assuming the
+#'   random effect is also included as a fixed effect). As such, a coefficient
+#'   close to 0 means that the participants' effect is the same as the
+#'   population-level effect (in other words, it is "in the norm"). If
+#'   `"total"`, it will return the sum of the random effect and its
+#'   corresponding fixed effects, which internally relies on the `coef()` method
+#'   (see `?coef.merMod`). Note that `type = "total"` yet does not return
+#'   uncertainty indices (such as SE and CI) for models from *lme4* or
+#'   *glmmTMB*, as the necessary information to compute them is not yet
+#'   available. However, for Bayesian models, it is possible to compute them.
+#' @param dispersion,test,diagnostic Arguments passed to
+#'    [parameters::model_parameters()] for Bayesian models. By default, it won't
+#'    return significance or diagnostic indices (as it is not typically very
+#'    useful).
+#' @param ... Other arguments passed to [parameters::model_parameters()].
 #'
 #' @details
 #' Unlike raw group means, BLUPs apply shrinkage: they are a compromise between
@@ -51,9 +55,23 @@
 #' # overall coefficients
 #' estimate_grouplevel(model, type = "total")
 #' @export
-estimate_grouplevel <- function(model, type = "random", ...) {
+estimate_grouplevel <- function(model, ...) {
+  UseMethod("estimate_grouplevel")
+}
+
+
+#' @rdname estimate_grouplevel
+#' @export
+estimate_grouplevel.default <- function(model,
+                                        type = "random",
+                                        ...) {
   # validate argument
   type <- insight::validate_argument(type, c("random", "total"))
+
+  # sanity check
+  if (is.null(insight::find_random(model))) {
+    insight::format_error("Model must be a mixed model with random effects.")
+  }
 
   # Extract params
   params <- parameters::model_parameters(
@@ -63,12 +81,15 @@ estimate_grouplevel <- function(model, type = "random", ...) {
     ...
   )
 
+  # get cleaned parameter names with additional information
+  clean_parameters <- attributes(params)$clean_parameters
+
[…]
