+++ tests/testthat/test-estimate_grouplevel.R
@@ -1,6 +1,11 @@
 skip_if(utils::packageVersion("insight") <= "1.1.0")
 skip_if(utils::packageVersion("parameters") <= "0.24.1")
 
+test_that("estimate_grouplevel - errors", {
+  m <- lm(mpg ~ gear, data = mtcars)
+  expect_error(estimate_grouplevel(m), regex = "Model must be a mixed")
+})
+
 test_that("estimate_grouplevel - lme4", {
   skip_if_not_installed("lme4")
   set.seed(333)
@@ -83,7 +88,7 @@ test_that("estimate_grouplevel - glmmTMB", {
   expect_named(out, c("Group", "Level", "Parameter", "Coefficient"))
 })
 
-test_that("estimate_grouplevel - Bayesian", {
+test_that("estimate_grouplevel - Bayesian, brms", {
   skip_on_cran()
   skip_if_not_installed("curl")
   skip_if_offline()
@@ -94,10 +99,10 @@ test_that("estimate_grouplevel - Bayesian", {
   skip_if(is.null(m))
 
   out <- estimate_grouplevel(m)
-  expect_identical(dim(out), c(6L, 7L))
-  expect_named(out, c("Group", "Level", "Parameter", "Median", "CI", "CI_low", "CI_high"))
+  expect_identical(dim(out), c(6L, 8L))
+  expect_named(out, c("Group", "Level", "Parameter", "Median", "MAD", "CI", "CI_low", "CI_high"))
 
-  out <- estimate_grouplevel(m, type = "total")
+  out <- estimate_grouplevel(m, type = "total", dispersion = FALSE)
   expect_identical(dim(out), c(6L, 7L))
   expect_named(out, c("Group", "Level", "Parameter", "Median", "CI", "CI_low", "CI_high"))
 
@@ -105,10 +110,32 @@ test_that("estimate_grouplevel - Bayesian", {
   skip_if(is.null(m))
 
   out <- estimate_grouplevel(m)
-  expect_identical(dim(out), c(12L, 8L))
-  expect_named(out, c("Component", "Group", "Level", "Parameter", "Median", "CI", "CI_low", "CI_high"))
+  expect_identical(dim(out), c(12L, 9L))
+  expect_named(out, c("Component", "Group", "Level", "Parameter", "Median", "MAD", "CI", "CI_low", "CI_high"))
 
   out <- estimate_grouplevel(m, type = "total")
-  expect_identical(dim(out), c(12L, 8L))
-  expect_named(out, c("Component", "Group", "Level", "Parameter", "Median", "CI", "CI_low", "CI_high"))
+  expect_identical(dim(out), c(12L, 9L))
+  expect_named(out, c("Component", "Group", "Level", "Parameter", "Median", "MAD", "CI", "CI_low", "CI_high"))
+})
+
+test_that("estimate_grouplevel - Bayesian, rstanarm", {
+  skip_on_cran()
+  skip_if_not_installed("curl")
+  skip_if_offline()
+  skip_if_not_installed("httr2")
+  skip_if_not_installed("rstanarm")
+
+  m <- insight::download_model("stanreg_merMod_1")
+  skip_if(is.null(m))
+
+  out <- estimate_grouplevel(m)
+  expect_identical(dim(out), c(4L, 8L))
+  expect_named(out, c("Group", "Level", "Parameter", "Median", "MAD", "CI", "CI_low", "CI_high"))
+
+  out <- estimate_grouplevel(m, dispersion = FALSE)
+  expect_identical(dim(out), c(4L, 7L))
+
+  out <- estimate_grouplevel(m, type = "total", dispersion = FALSE)
+  expect_identical(dim(out), c(3L, 4L))
+  expect_named(out, c("Group", "Level", "Parameter", "Coefficient"))
 })
