### 2025-03-12 58bb861
Deploying to gh-pages from @ easystats/modelbased@2582483b8b94765a185b7628fc807fbc6a1279a6 🚀
