### 2025-03-13 74599eb
Deploying to gh-pages from @ easystats/modelbased@8c78ce0b3a2c35db6ccc7b2cceb350a3f2dd942d 🚀
