+++ tests/testthat/_snaps/windows/ordinal.md
@@ -49,38 +49,38 @@
     Output
       Estimated Marginal Means
       
-      Sepal.Width | Response   | Median |       95% CI |   pd |          ROPE | % in ROPE
+      Sepal.Width |   Response | Median |       95% CI |   pd |          ROPE | % in ROPE
       -----------------------------------------------------------------------------------
-      2.00        | setosa     |   0.00 | [0.00, 0.01] | 100% | [-0.10, 0.10] |      100%
-      2.00        | versicolor |   0.86 | [0.65, 0.95] | 100% | [-0.10, 0.10] |        0%
-      2.00        | virginica  |   0.14 | [0.05, 0.34] | 100% | [-0.10, 0.10] |    28.00%
-      2.27        | setosa     |   0.00 | [0.00, 0.02] | 100% | [-0.10, 0.10] |      100%
-      2.27        | versicolor |   0.78 | [0.60, 0.90] | 100% | [-0.10, 0.10] |        0%
-      2.27        | virginica  |   0.21 | [0.10, 0.40] | 100% | [-0.10, 0.10] |        0%
-      2.53        | setosa     |   0.02 | [0.01, 0.06] | 100% | [-0.10, 0.10] |      100%
-      2.53        | versicolor |   0.66 | [0.52, 0.78] | 100% | [-0.10, 0.10] |        0%
-      2.53        | virginica  |   0.32 | [0.20, 0.45] | 100% | [-0.10, 0.10] |        0%
-      2.80        | setosa     |   0.09 | [0.04, 0.16] | 100% | [-0.10, 0.10] |    65.72%
-      2.80        | versicolor |   0.49 | [0.39, 0.59] | 100% | [-0.10, 0.10] |        0%
-      2.80        | virginica  |   0.41 | [0.32, 0.52] | 100% | [-0.10, 0.10] |        0%
-      3.07        | setosa     |   0.27 | [0.19, 0.36] | 100% | [-0.10, 0.10] |        0%
-      3.07        | versicolor |   0.30 | [0.21, 0.39] | 100% | [-0.10, 0.10] |        0%
-      3.07        | virginica  |   0.43 | [0.33, 0.53] | 100% | [-0.10, 0.10] |        0%
-      3.33        | setosa     |   0.57 | [0.47, 0.68] | 100% | [-0.10, 0.10] |        0%
-      3.33        | versicolor |   0.12 | [0.06, 0.21] | 100% | [-0.10, 0.10] |    29.76%
-      3.33        | virginica  |   0.30 | [0.20, 0.40] | 100% | [-0.10, 0.10] |        0%
-      3.60        | setosa     |   0.82 | [0.68, 0.91] | 100% | [-0.10, 0.10] |        0%
-      3.60        | versicolor |   0.03 | [0.01, 0.08] | 100% | [-0.10, 0.10] |      100%
-      3.60        | virginica  |   0.14 | [0.07, 0.26] | 100% | [-0.10, 0.10] |    17.37%
-      3.87        | setosa     |   0.94 | [0.83, 0.98] | 100% | [-0.10, 0.10] |        0%
-      3.87        | versicolor |   0.01 | [0.00, 0.03] | 100% | [-0.10, 0.10] |      100%
-      3.87        | virginica  |   0.05 | [0.02, 0.15] | 100% | [-0.10, 0.10] |    89.59%
-      4.13        | setosa     |   0.98 | [0.91, 1.00] | 100% | [-0.10, 0.10] |        0%
-      4.13        | versicolor |   0.00 | [0.00, 0.01] | 100% | [-0.10, 0.10] |      100%
[…]
