+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.10.0.4
+Version: 0.10.0.3
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -102,4 +102,4 @@ Roxygen: list(markdown = TRUE)
 Config/Needs/check: stan-dev/cmdstanr
 Config/Needs/website: easystats/easystatstemplate
 LazyData: true
-Remotes: easystats/parameters
+Remotes: easystats/bayestestR, easystats/parameters
