### 2025-03-14 f8e8789
Inhouse-processing of parameter-tidying (#447)
* Inhouse-processing of parameter-tidying

* no nested if-sztatement

* Update DESCRIPTION

* Update DESCRIPTION

* Update ordinal.md

* Update test-estimate_grouplevel.R

* Update DESCRIPTION
