+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 3.6.3
 pkgdown: 2.1.1.9000
-pkgdown_sha: 7cf401f44a970ea13744fcdd883248552c68ff9d
+pkgdown_sha: d5a48705424d7745e1e62354505ca70091931e61
 articles:
   derivatives: derivatives.html
   estimate_contrasts: estimate_contrasts.html
@@ -19,7 +19,7 @@ articles:
   practical_intersectionality: practical_intersectionality.html
   visualisation_matrix: visualisation_matrix.html
   workflow_modelbased: workflow_modelbased.html
-last_built: 2025-03-13T07:29Z
+last_built: 2025-03-14T12:40Z
 urls:
   reference: https://easystats.github.io/modelbased/reference
   article: https://easystats.github.io/modelbased/articles
