### 2025-03-17 516cecd
Deploying to gh-pages from @ easystats/modelbased@1c18bce87ad5a0b90c93ed105d1ba447dce739dd 🚀
