### 2025-03-18 #7894 @spholmes
👋 @vincentarelbundock , I wonder whether you would be able to review the JOSS submission about the CRAN
package modelbased, I saw on bluesky that you  already use it. The review is checklist based and not too onerous for knowledgeable github practioners...
Thanks,
Susan
