### 2025-03-18 #7894 @spholmes
👋  @bbolker
I wonder whether you have a moment to review the modelbased submission to JOSS, a quick No being the second best answer as I think we can find someone interested in this, I know you are probably overloaded with requests, but I thought of
you recently and all my Canadian friends, hi from 🇵‍🇹 where I now retired....warm wishes, Susan
