### 2025-03-20 1e8b0d0
Refactor, DRY
