+++ NEWS.md
@@ -12,6 +12,9 @@
 
 * Fixed issues in `estimate_grouplevel()` for models from package *rstanarm*.
 
+* Fixed issues in calculating correct confidence intervals (and possibly p-values)
+  for pooling functions `pool_parameters()` and `pool_predictions()`.
+
 # modelbased 0.10.0
 
 ## Breaking Changes
