### 2025-03-20 9761b40
allow pooling slopes
