+++ tests/testthat/test-mice.R
@@ -38,3 +38,34 @@ test_that("pool_contrasts", {
   expect_equal(out$CI_low, c(-12.31066, -18.92406, -11.94194), tolerance = 1e-3)
   expect_equal(out$p, c(0.14926, 0.17899, 0.55449), tolerance = 1e-3)
 })
+
+
+
+test_that("pool_slopes", {
+  set.seed(123)
+  data("nhanes2", package = "mice")
+  imp <- mice::mice(nhanes2, printFlag = FALSE)
+  slopes <- lapply(1:5, function(i) {
+    m <- lm(bmi ~ age + hyp + chl, data = mice::complete(imp, action = i))
+    estimate_slopes(m, "chl")
+  })
+  out <- pool_slopes(slopes)
+  expect_equal(out$Slope, 0.05666, tolerance = 1e-3)
+  expect_equal(out$CI_low, 0.00395, tolerance = 1e-3)
+  expect_named(
+    out,
+    c("Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
+  )
+
+  slopes <- lapply(1:5, function(i) {
+    m <- lm(bmi ~ age + hyp + chl, data = mice::complete(imp, action = i))
+    estimate_slopes(m, "age")
+  })
+  out <- pool_slopes(slopes)
+  expect_equal(out$Slope, c(-4.6464, -6.70639), tolerance = 1e-3)
+  expect_equal(out$CI_low, c(-9.36657, -14.23085), tolerance = 1e-3)
+  expect_named(
+    out,
+    c("Comparison", "Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
+  )
+})
