+++ tests/testthat/test-mice.R
@@ -11,8 +11,8 @@ test_that("pool_predictions", {
     estimate_means(m, "age")
   })
   out <- pool_predictions(predictions)
-  expect_equal(out$Mean, c(29.84661, 25.20021, 23.14022), tolerance = 1e-3)
-  expect_equal(out$CI_low, c(2.10117, 3.44548, -5.79522), tolerance = 1e-3)
+  expect_equal(out$Mean, c(29.84661, 25.20021, 23.14022), tolerance = 1e-2)
+  expect_equal(out$CI_low, c(2.10117, 3.44548, -5.79522), tolerance = 1e-2)
 
   # transformed response
   predictions <- lapply(1:5, function(i) {
@@ -20,8 +20,8 @@ test_that("pool_predictions", {
     estimate_means(m, "age")
   })
   out <- pool_predictions(predictions, transform = TRUE)
-  expect_equal(out$Mean, c(29.67473, 24.99382, 23.19148), tolerance = 1e-3)
-  expect_equal(out$CI_low, c(10.58962, 11.13011, 7.43196), tolerance = 1e-3)
+  expect_equal(out$Mean, c(29.67473, 24.99382, 23.19148), tolerance = 1e-2)
+  expect_equal(out$CI_low, c(10.58962, 11.13011, 7.43196), tolerance = 1e-2)
 })
 
 
@@ -34,9 +34,9 @@ test_that("pool_contrasts", {
     estimate_contrasts(m, "age")
   })
   out <- pool_contrasts(comparisons)
-  expect_equal(out$Difference, c(-4.6464, -6.70639, -2.05999), tolerance = 1e-3)
-  expect_equal(out$CI_low, c(-12.31066, -18.92406, -11.94194), tolerance = 1e-3)
-  expect_equal(out$p, c(0.14926, 0.17899, 0.55449), tolerance = 1e-3)
+  expect_equal(out$Difference, c(-4.6464, -6.70639, -2.05999), tolerance = 1e-2)
+  expect_equal(out$CI_low, c(-12.31066, -18.92406, -11.94194), tolerance = 1e-2)
+  expect_equal(out$p, c(0.14926, 0.17899, 0.55449), tolerance = 1e-2)
 })
 
 
@@ -50,8 +50,8 @@ test_that("pool_slopes", {
     estimate_slopes(m, "chl")
   })
   out <- pool_slopes(slopes)
-  expect_equal(out$Slope, 0.05666, tolerance = 1e-3)
-  expect_equal(out$CI_low, 0.00395, tolerance = 1e-3)
+  expect_equal(out$Slope, 0.05666, tolerance = 1e-2)
+  expect_equal(out$CI_low, 0.00395, tolerance = 1e-2)
   expect_named(
     out,
     c("Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
@@ -62,8 +62,8 @@ test_that("pool_slopes", {
     estimate_slopes(m, "age")
   })
   out <- pool_slopes(slopes)
-  expect_equal(out$Slope, c(-4.6464, -6.70639), tolerance = 1e-3)
-  expect_equal(out$CI_low, c(-9.36657, -14.23085), tolerance = 1e-3)
+  expect_equal(out$Slope, c(-4.6464, -6.70639), tolerance = 1e-2)
+  expect_equal(out$CI_low, c(-9.36657, -14.23085), tolerance = 1e-2)
   expect_named(
     out,
     c("Comparison", "Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
