+++ tests/testthat/test-mice.R
@@ -53,7 +53,7 @@ test_that("pool_slopes", {
   expect_equal(out$CI_low, 0.00395, tolerance = 1e-2)
   expect_named(
     out,
-    c("Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
+    c("Slope", "SE", "CI_low", "CI_high", "t", "p", "df")
   )
 
   slopes <- lapply(1:5, function(i) {
@@ -65,6 +65,6 @@ test_that("pool_slopes", {
   expect_equal(out$CI_low, c(-9.36657, -14.23085), tolerance = 1e-2)
   expect_named(
     out,
-    c("Comparison", "Slope", "SE", "CI_low", "CI_high", "t", "p",  "df")
+    c("Comparison", "Slope", "SE", "CI_low", "CI_high", "t", "p", "df")
   )
 })
