### 2025-03-20 34a057f
Refactor, DRY (#448)
* Refactor, DRY

* fix

* fix

* fix

* news, test

* allow pooling slopes

* news

* decrease tolerance

* styler

* styler
