+++ R/pool.R
@@ -40,7 +40,6 @@ pool_contrasts <- function(x, ...) {
 
   # preparation ----
 
-  len <- length(x)
   ci <- attributes(x[[1]])$ci
   dof <- x[[1]]$df
 
@@ -48,40 +47,11 @@ pool_contrasts <- function(x, ...) {
     dof <- Inf
   }
 
-  # pool predictions -----
-
-  pooled_comparisons <- original_x[[1]]
-  pooled_comparisons$SE <- NA
-  n_rows <- nrow(original_x[[1]])
-
-  for (i in 1:n_rows) {
-    # pooled estimate
-    pooled_comp <- unlist(lapply(original_x, function(j) j[[estimate_name]][i]), use.names = FALSE)
-    pooled_comparisons[[estimate_name]][i] <- mean(pooled_comp, na.rm = TRUE)
-
-    # pooled standard error
-    pooled_se <- unlist(lapply(original_x, function(j) j$SE[i]), use.names = FALSE)
-    ubar <- mean(pooled_se^2, na.rm = TRUE)
-    tmp <- ubar + (1 + 1 / len) * stats::var(pooled_comp)
-    pooled_comparisons$SE[i] <- sqrt(tmp)
-  }
-
-  # pooled degrees of freedom for t-statistics
-  pooled_df <- .barnad_rubin(
-    m = nrow(pooled_comparisons),
-    b = stats::var(pooled_comparisons[[estimate_name]]),
-    t = pooled_comparisons$SE^2,
-    dfcom = dof
-  )
+  # pool contrasts -----
+  pooled_comparisons <- .pool_estimates(original_x, estimate_name, original_x[[1]])
 
   # confidence intervals ----
-  alpha <- (1 + ci) / 2
-  fac <- stats::qt(alpha, df = dof)
-  pooled_comparisons$CI_low <- pooled_comparisons[[estimate_name]] - fac * pooled_comparisons$SE
-  pooled_comparisons$CI_high <- pooled_comparisons[[estimate_name]] + fac * pooled_comparisons$SE
-
-  # udpate df ----
-  pooled_comparisons$df <- pooled_df
+  pooled_comparisons <- .pool_ci_and_stats(pooled_comparisons, estimate_name, ci, dof)
 
   attributes(pooled_comparisons) <- utils::modifyList(attributes(original_x[[1]]), attributes(pooled_comparisons))
   pooled_comparisons
@@ -93,31 +63,45 @@ pool_contrasts <- function(x, ...) {
 #' This function "pools" (i.e. combines) multiple `estimate_means` objects, in
 #' a similar fashion as [`mice::pool()`].
 #'
-#' @param x A list of `estimate_means` objects, as returned by [`estimate_means()`],
-#' or `estimate_predicted`, as returned by [`estimate_relation()`] and related
-#' functions.
+#' @param x A list of `estimate_means` objects, as returned by
+#' [`estimate_means()`], or `estimate_predicted` objects, as returned by
+#' [`estimate_relation()`] and related functions. For `pool_slopes()`, must be
+#' a list of `estimate_slopes` objects, as returned by [`estimate_slopes()`].
 #' @param ... Currently not used.
 #' @inheritParams estimate_means
 #'
 #' @details Averaging of parameters follows Rubin's rules (*Rubin, 1987, p. 76*).
-#' Pooling is applied to the predicted values on the scale of the *linear predictor*,
-#' not on the response scale, in order to have accurate pooled estimates and
-#' standard errors. The final pooled predicted values are then transformed to
-#' the response scale, using [`insight::link_inverse()`].
[…]
