+++ NAMESPACE
@@ -82,6 +82,7 @@ export(get_marginalmeans)
 export(get_marginaltrends)
 export(pool_contrasts)
 export(pool_predictions)
+export(pool_slopes)
 export(print_html)
 export(print_md)
 export(reshape_grouplevel)

[…]
