### 2025-03-21 8cf44f0
Deploying to gh-pages from @ easystats/modelbased@7b7c02f2d9b3ed224ad2610f15b60a9f1534f1d6 🚀
