### 2025-03-22 e5bbf1e
better formatting for emmeans
