+++ R/estimate_means.R
@@ -142,6 +142,32 @@
 #'
 #' @return A data frame of estimated marginal means.
 #'
+#' @section Predictions and contrasts at meaningful values (data grids):
+#'
+#' To define representative values for focal predictors (specified in `by`,
+#' `contrast`, and `trend`), you can use several methods. These values are
+#' internally generated by `insight::get_datagrid()`, so consult its
+#' documentation for more details.
+#'
+#' * You can directly specify values as strings or lists for `by`, `contrast`,
+#'   and `trend`.
+#'   * For numeric predictors, use examples like:
+#'     * `by = "gear=c(4,8)"`
+#'     * `by = list(gear = c(4, 8))`
+#' * For factor or character predictors, use:
+#'   * `by = "Species=c('setosa', 'virginica')"`
+#'   * `by = list(Species = c('setosa', 'virginica'))`
+#' * You can use "shortcuts" within square brackets, such as:
+#'    * `by = "Sepal.Width=[sd]"`
+#'    * `by = "Sepal.Width=[1.5,2.5,3.5]"`
+#' * For numeric predictors, `length` and `range` control the number and type of
+#'   representative values:
+#'   * `length` determines how many equally spaced values are generated.
+#'   * `range` specifies the type of values, like `"range"` or `"sd"`.
+#'   * `length` and `range` apply to all numeric predictors.
+#'   * If you have multiple numeric predictors, `length` and `range` can accept
+#'     multiple elements, one for each predictor.
+#'
 #' @section Global Options to Customize Estimation of Marginal Means:
 #'
 #' - `modelbased_backend`: `options(modelbased_backend = <string>)` will set a
