+++ man/estimate_means.Rd
@@ -225,6 +225,45 @@ these levels (using \code{\link[=estimate_contrasts]{estimate_contrasts()}}). Fi
 the effect of x averaged over all conditions, or instead within each
 condition (\code{using [estimate_slopes]}).
 }
+\section{Predictions and contrasts at meaningful values (data grids)}{
+
+
+To define representative values for focal predictors (specified in \code{by},
+\code{contrast}, and \code{trend}), you can use several methods. These values are
+internally generated by \code{insight::get_datagrid()}, so consult its
+documentation for more details.
+\itemize{
+\item You can directly specify values as strings or lists for \code{by}, \code{contrast},
+and \code{trend}.
+\itemize{
+\item For numeric predictors, use examples like:
+\itemize{
+\item \code{by = "gear=c(4,8)"}
+\item \code{by = list(gear = c(4, 8))}
+}
+}
+\item For factor or character predictors, use:
+\itemize{
+\item \code{by = "Species=c('setosa', 'virginica')"}
+\item \code{by = list(Species = c('setosa', 'virginica'))}
+}
+\item You can use "shortcuts" within square brackets, such as:
+\itemize{
+\item \code{by = "Sepal.Width=[sd]"}
+\item \code{by = "Sepal.Width=[1.5,2.5,3.5]"}
+}
+\item For numeric predictors, \code{length} and \code{range} control the number and type of
+representative values:
+\itemize{
+\item \code{length} determines how many equally spaced values are generated.
+\item \code{range} specifies the type of values, like \code{"range"} or \code{"sd"}.
+\item \code{length} and \code{range} apply to all numeric predictors.
+\item If you have multiple numeric predictors, \code{length} and \code{range} can accept
+multiple elements, one for each predictor.
+}
+}
+}
+
 \section{Global Options to Customize Estimation of Marginal Means}{
 
 \itemize{
