+++ R/estimate_means.R
@@ -151,15 +151,12 @@
 #'
 #' * You can directly specify values as strings or lists for `by`, `contrast`,
 #'   and `trend`.
-#'   * For numeric focal predictors, use examples like:
-#'     * `by = "gear=c(4,8)"`
-#'     * `by = list(gear = c(4, 8))`
-#' * For factor or character predictors, use:
-#'   * `by = "Species=c('setosa', 'virginica')"`
-#'   * `by = list(Species = c('setosa', 'virginica'))`
-#' * You can use "shortcuts" within square brackets, such as:
-#'    * `by = "Sepal.Width=[sd]"`
-#'    * `by = "Sepal.Width=[1.5,2.5,3.5]"`
+#'   * For numeric focal predictors, use examples like `by = "gear=c(4,8)"` or
+#'     `by = list(gear = c(4, 8))`
+#'   * For factor or character predictors, use `by = "Species=c('setosa', 'virginica')"`
+#'     or `by = list(Species = c('setosa', 'virginica'))`
+#' * You can use "shortcuts" within square brackets, such as `by = "Sepal.Width=[sd]"`
+#'   or `by = "Sepal.Width=[1.5,2.5,3.5]"`
 #' * For numeric focal predictors, `length` and `range` control the number and
 #'   type of representative values:
 #'   * `length` determines how many equally spaced values are generated.
