### 2025-03-22 5949cbd
mention speed issues
