### 2025-03-22 dec1b0e
minor docs: say "Significantly slower" (as it can be more than 10 times)
