### 2025-03-22 0ab3384
Backend differences (#451)
* Backend differences
Fixes #450

* suppress notes

* better formatting for emmeans

* docs

* docs

* reference to vignette

* revise

* mention speed issues

* no linebreak

* table footer shows % points

* Update estimate_contrasts.md

* typo

* docs

* minor docs: say "Significantly slower" (as it can be more than 10 times)

* docs

* docs

---------

Co-authored-by: DominiqueMakowski <dom.mak19@gmail.com>
