+++ man/estimate_contrasts.Rd
@@ -34,7 +34,9 @@ to functions from the \strong{emmeans} or \strong{marginaleffects} package, or t
 Bayesian models via \code{\link[bayestestR:describe_posterior]{bayestestR::describe_posterior()}}. Examples:
 \itemize{
 \item \code{insight::get_datagrid()}: Argument such as \code{length}, \code{digits} or \code{range}
-can be used to control the (number of) representative values.
+can be used to control the (number of) representative values. For integer
+variables, \code{protect_integers} modulates whether these should also be
+treated as numerics, i.e. values can have fractions or not.
 \item \strong{marginaleffects}: Internally used functions are \code{avg_predictions()} for
 means and contrasts, and \code{avg_slope()} for slopes. Therefore, arguments for
 instance like \code{vcov}, \code{equivalence}, \code{df}, \code{slope} or even \code{newdata} can be
@@ -174,7 +176,7 @@ This is also the option that should be used for \strong{G-computation}
 }
 
 You can set a default option for the \code{estimate} argument via \code{options()},
-e.g. \code{options(modelbased_estimate = "average")}}
+e.g. \code{options(modelbased_estimate="average")}}
 
 \item{p_adjust}{The p-values adjustment method for frequentist multiple
 comparisons. Can be one of \code{"none"} (default), \code{"hochberg"}, \code{"hommel"},
@@ -213,14 +215,19 @@ using \code{effectsize = "boot"}. One of \code{"unstandardized"}, \code{"cohens.
 
 \item{backend}{Whether to use \code{"marginaleffects"} (default) or \code{"emmeans"} as
 a backend. Results are usually very similar. The major difference will be
-found for mixed models, where \code{backend = "marginaleffects"} will also average
+found for mixed models, where \code{backend="marginaleffects"} will also average
 across random effects levels, producing "marginal predictions" (instead of
 "conditional predictions", see Heiss 2022).
 
+Another difference is that \code{backend="marginaleffects"} will be slower than
+\code{backend="emmeans"}. For most models, this difference is negligible. However,
+in particular complex models or large data sets fitted with \emph{glmmTMB} can be
+significantly slower.
+
 You can set a default backend via \code{options()}, e.g. use
-\code{options(modelbased_backend = "emmeans")} to use the \strong{emmeans} package or
-\code{options(modelbased_backend = "marginaleffects")} to set \strong{marginaleffects}
-as default backend.}
+\code{options(modelbased_backend="emmeans")} to use the \strong{emmeans} package or
+\code{options(modelbased_backend="marginaleffects")} to set \strong{marginaleffects} as
+default backend.}
 
 \item{verbose}{Use \code{FALSE} to silence messages and warnings.}
 }
@@ -318,6 +325,43 @@ covariates, but creates an appropriate scale for standardizing the response.
 
 \code{effectsize = "boot"} uses bootstrapping (defaults to a low value of
 200) through \link[bootES:bootES]{bootES::bootES}. Adjusts for contrasts, but not for covariates.
+}
+
+\section{Predictions and contrasts at meaningful values (data grids)}{
+
+
+To define representative values for focal predictors (specified in \code{by},
+\code{contrast}, and \code{trend}), you can use several methods. These values are
+internally generated by \code{insight::get_datagrid()}, so consult its
+documentation for more details.
+\itemize{
+\item You can directly specify values as strings or lists for \code{by}, \code{contrast},
+and \code{trend}.
+\itemize{
+\item For numeric focal predictors, use examples like \code{by="gear=c(4,8)"} or
+\code{by=list(gear=c(4,8))}
+\item For factor or character predictors, use \code{by="Species=c('setosa','virginica')"}
+or \code{by=list(Species=c('setosa','virginica'))}
+}
[…]
