### 2025-03-22 9067b97
Deploying to gh-pages from @ easystats/modelbased@66a5941ddc1ec6b61b0cd8760b074cfde2dc39bf 🚀
