### 2025-03-23 a0fb6c5
Deploying to gh-pages from @ easystats/modelbased@309d2dac045642a5716e910bbcc6247d406e5e1d 🚀
