+++ R/estimate_means.R
@@ -12,17 +12,15 @@
 #' included here will be collapsed and "averaged" over (the effect will be
 #' estimated across them). `by` can be a character (vector) naming the focal
 #' predictors, optionally including representative values or levels at which
-#' focal predictors are evaluated (e.g., `by="x=c(1,2)"`). When `estimate` is
-#' *not* `"average"`, the `by` argument is used to create a "reference grid" or
-#' "data grid" with representative values for the focal predictors. In this
+#' focal predictors are evaluated (e.g., `by = "x = c(1, 2)"`). When `estimate`
+#' is *not* `"average"`, the `by` argument is used to create a "reference grid"
+#' or "data grid" with representative values for the focal predictors. In this
 #' case, `by` can also be list of named elements. See details in
 #' [`insight::get_datagrid()`] to learn more about how to create data grids for
 #' predictors of interest.
 #' @param predict Is passed to the `type` argument in `emmeans::emmeans()` (when
 #' `backend = "emmeans"`) or in `marginaleffects::avg_predictions()` (when
-#' `backend = "marginaleffects"`). For emmeans, see also
-#' [this vignette](https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html).
-#' Valid options for `predict` are:
+#' `backend = "marginaleffects"`). Valid options for `predict` are:
 #'
 #' * `backend = "marginaleffects"`: `predict` can be `"response"`, `"link"`,
 #'   `"inverse_link"` or any valid `type` option supported by model's class
@@ -36,7 +34,8 @@
 #'   response scale.
 #' * `backend = "emmeans"`: `predict` can be `"response"`, `"link"`, `"mu"`,
 #'   `"unlink"`, or `"log"`. If `predict = NULL` (default), the most appropriate
-#'   transformation is selected (which usually is `"response"`).
+#'   transformation is selected (which usually is `"response"`). See also
+#'   [this vignette](https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html).
 #'
 #' `"link"` will leave the values on scale of the linear predictors.
 #' `"response"` (or `NULL`) will transform them on scale of the response
@@ -90,21 +89,21 @@
 #'   (_Chatton and Rohrer 2024_).
 #'
 #' You can set a default option for the `estimate` argument via `options()`,
-#' e.g. `options(modelbased_estimate="average")`
+#' e.g. `options(modelbased_estimate = "average")`
 #' @param backend Whether to use `"marginaleffects"` (default) or `"emmeans"` as
 #' a backend. Results are usually very similar. The major difference will be
-#' found for mixed models, where `backend="marginaleffects"` will also average
+#' found for mixed models, where `backend = "marginaleffects"` will also average
 #' across random effects levels, producing "marginal predictions" (instead of
 #' "conditional predictions", see Heiss 2022).
 #'
-#' Another difference is that `backend="marginaleffects"` will be slower than
-#' `backend="emmeans"`. For most models, this difference is negligible. However,
+#' Another difference is that `backend = "marginaleffects"` will be slower than
+#' `backend = "emmeans"`. For most models, this difference is negligible. However,
 #' in particular complex models or large data sets fitted with *glmmTMB* can be
 #' significantly slower.
 #'
 #' You can set a default backend via `options()`, e.g. use
-#' `options(modelbased_backend="emmeans")` to use the **emmeans** package or
-#' `options(modelbased_backend="marginaleffects")` to set **marginaleffects** as
+#' `options(modelbased_backend = "emmeans")` to use the **emmeans** package or
[…]
