+++ vignettes/plotting.Rmd
@@ -14,6 +14,8 @@ knitr::opts_chunk$set(
   comment = "#>",
   dev = "png",
   out.width = "100%",
+  fig.width = 7,
+  fig.height = 4,
   dpi = 300,
   message = FALSE,
   warning = FALSE,
@@ -34,7 +36,13 @@ if (getRversion() < "4.1.0") {
 
 This vignette provides a quick overview with different examples that show how to plot estimated marginal means.
 
+In summary, you can use the `length` and `range` arguments in `estimate_means()` (which are passed to [`insight::get_datagrid()](https://easystats.github.io/insight/reference/get_datagrid.html)), as well as directly specifying meaningful values in the `by` argument, which are also used to create a data grid, to control the plot-appearance.
+
+See also the [vignette on data grids](https://easystats.github.io/modelbased/articles/visualisation_matrix.html).
+
 ## One predictor - categorical
+
+The simplest case is possibly plotting one categorical predictor. Predicted values for each level and its confidence intervals are shown.
 
 ```{r}
 library(modelbased)
@@ -47,11 +55,15 @@ estimate_means(m, "c172code") |> plot()
 
 ## One predictor - numeric
 
+For numeric predictors, the range of predictions at different values of the focal predictor are plotted, the uncertainty is displayed as confidence band.
+
 ```{r}
 estimate_means(m, "barthtot") |> plot()
 ```
 
 ## Two predictors - categorical
+
+For two categorical predictors, the first focal predictors is plotted along the x-asis, while the levels of the second predictor are mapped to different colors.
 
 ```{r}
 m <- lm(neg_c_7 ~ e16sex * c172code + e42dep, data = efc)
@@ -60,79 +72,92 @@ estimate_means(m, c("e16sex", "c172code")) |> plot()
 
 ## Two predictors - numeric * categorical
 
+For two predictors, where the first ist numeric and the second categorical, range of predictions including confidence bands are shown, with the different levels of the second (categorical) predictor mapped to colors again.
+
 ```{r}
 m <- lm(neg_c_7 ~ barthtot * c172code + e42dep, data = efc)
 estimate_means(m, c("barthtot", "c172code")) |> plot()
 ```
 
-## Two predictors - categorical * numeric
+In general, plots can be further modified using functions from the **ggplot2** package. Thereby, other themes, color scales, facetting and so on, can be applies.
 
 ```{r}
+library(ggplot2)
+estimate_means(m, c("barthtot", "c172code")) |>
+  plot() +
+  see::theme_modern(show.ticks = TRUE)
+
+estimate_means(m, c("barthtot", "c172code")) |>
+  plot() +
+  facet_grid(~c172code)
+
+estimate_means(m, c("barthtot", "c172code")) |>
+  plot() +
+  scale_color_brewer(palette = "Dark2") +
+  scale_fill_brewer(palette = "Dark2")
+```
+
+## Two predictors - categorical * numeric
+
+If the numeric predictor is the _second_ focal term, its values are still mapped to colors, however, by default to a continuous (gradient) scale, because a range of representative values for that numeric predictor is used by default.
+
+Focal predictors specified in `estimate_means()` are passed to `insight::get_datagrid()`. If not specified otherwise, representative values for numeric predictors are evenly distributed from the minimum to the maximum, with a total number of `length` values covering that range.
+
[…]
