### 2025-03-26 e665a0e
Draft plotting vignette (#449)
* Draft

* vignette

* finalize

* update pkgdown

* typos

* final
