+++ vignettes/plotting.Rmd
@@ -1,0 +1,178 @@
+---
+title: "Plotting estimated marginal means"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Plotting estimated marginal means}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+bibliography: bibliography.bib
+---
+
+```{r set-options, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  dev = "png",
+  out.width = "100%",
+  fig.width = 7,
+  fig.height = 4,
+  dpi = 300,
+  message = FALSE,
+  warning = FALSE,
+  package.startup.message = FALSE
+)
+
+options(modelbased_join_dots = FALSE)
+options(modelbased_select = "minimal")
+
+pkgs <- c("marginaleffects", "ggplot2", "see")
+if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
+  knitr::opts_chunk$set(eval = FALSE)
+}
+if (getRversion() < "4.1.0") {
+  knitr::opts_chunk$set(eval = FALSE)
+}
+```
+
+This vignette provides a quick overview with different examples that show how to plot estimated marginal means.
+
+In summary, you can use the `length` and `range` arguments in `estimate_means()` (which are passed to [`insight::get_datagrid()](https://easystats.github.io/insight/reference/get_datagrid.html)), as well as directly specifying meaningful values in the `by` argument, which are also used to create a data grid, to control the plot-appearance. See also the [vignette on data grids](https://easystats.github.io/modelbased/articles/visualisation_matrix.html).
+
+Although the **modelbased** package does not focus on publication-ready plots, the default plots can already be used directly. Furthermore, a few modifications are already applies, like a percentage-scale for logistic regression models, or using variable labels for _labelled data_.
+
+## One predictor - categorical
+
+The simplest case is possibly plotting one categorical predictor. Predicted values for each level and its confidence intervals are shown.
+
+```{r}
+library(modelbased)
+data(efc, package = "modelbased")
+efc <- datawizard::to_factor(efc, c("e16sex", "c172code", "e42dep"))
+
+m <- lm(neg_c_7 ~ e16sex + c172code + barthtot, data = efc)
+estimate_means(m, "c172code") |> plot()
+```
+
+## One predictor - numeric
+
+For numeric predictors, the range of predictions at different values of the focal predictor are plotted, the uncertainty is displayed as confidence band.
+
+```{r}
+estimate_means(m, "barthtot") |> plot()
+```
+
+## Two predictors - categorical
+
+For two categorical predictors, the first focal predictors is plotted along the x-axis, while the levels of the second predictor are mapped to different colors.
+
+```{r}
+m <- lm(neg_c_7 ~ e16sex * c172code + e42dep, data = efc)
+estimate_means(m, c("e16sex", "c172code")) |> plot()
+```
+
+## Two predictors - numeric * categorical
+
+For two predictors, where the first is numeric and the second categorical, range of predictions including confidence bands are shown, with the different levels of the second (categorical) predictor mapped to colors again.
+
+```{r}
+m <- lm(neg_c_7 ~ barthtot * c172code + e42dep, data = efc)
+estimate_means(m, c("barthtot", "c172code")) |> plot()
+```
+
+In general, plots can be further modified using functions from the **ggplot2** package. Thereby, other themes, color scales, faceting and so on, can be applies.
+
+```{r}
+library(ggplot2)
+estimate_means(m, c("barthtot", "c172code")) |>
+  plot() +
+  see::theme_modern(show.ticks = TRUE)
+
+estimate_means(m, c("barthtot", "c172code")) |>
+  plot() +
[…]
