### 2025-03-26 2eb4cfd
Deploying to gh-pages from @ easystats/modelbased@e665a0edb58206ecf6e15da529dd87a37f647a5a 🚀
