+++ pkgdown.yml
@@ -15,11 +15,12 @@ articles:
   introduction_comparisons_4: introduction_comparisons_4.html
   modelisation_approach: modelisation_approach.html
   overview_of_vignettes: overview_of_vignettes.html
+  plotting: plotting.html
   practical_causality: practical_causality.html
   practical_intersectionality: practical_intersectionality.html
   visualisation_matrix: visualisation_matrix.html
   workflow_modelbased: workflow_modelbased.html
-last_built: 2025-03-23T21:17Z
+last_built: 2025-03-26T12:17Z
 urls:
   reference: https://easystats.github.io/modelbased/reference
   article: https://easystats.github.io/modelbased/articles
