+++ NEWS.md
@@ -22,6 +22,8 @@
 * Fixed issues in calculating correct confidence intervals (and possibly p-values)
   for pooling functions `pool_parameters()` and `pool_predictions()`.
 
+* Fixed issue with wrong y-axis label for plots from `estimate_slopes()`.
+
 # modelbased 0.10.0
 
 ## Breaking Changes
