### 2025-03-28 ba0a7f0
Question about the Johnson_Neyman plot axis labeling.. (#453)
* Question about the Johnson_Neyman plot axis labeling..
Fixes #452

* wordlist
