### 2025-03-28 264833e
Deploying to gh-pages from @ easystats/modelbased@ba0a7f0a2280f2d89c6499bda8b48bd6e5b869dc 🚀
