### 2025-03-30 67a617c
Revise docs, remove deprecated arg
