+++ R/get_marginalcontrasts.R
@@ -73,7 +73,7 @@ get_marginalcontrasts <- function(model,
 
   # if first focal term is numeric, we contrast slopes, but slopes only for
   # # numerics with many values, not for binary or likert-alike
-  if (is.numeric(model_data[[first_focal]]) && !.is_likert(model_data[[first_focal]]) && !first_focal %in% on_the_fly_factors) { # nolint
+  if (is.numeric(model_data[[first_focal]]) && !.is_likert(model_data[[first_focal]], ...) && !first_focal %in% on_the_fly_factors) { # nolint
     # sanity check - contrast for slopes only makes sense when we have a "by" argument
     if (is.null(my_args$by)) {
       insight::format_error("Please specify the `by` argument to calculate contrasts of slopes.") # nolint
