+++ R/estimate_means.R
@@ -37,31 +37,8 @@
 #'   transformation is selected (which usually is `"response"`). See also
 #'   [this vignette](https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html).
 #'
-#' The `"link"` option does not apply to all models, and usually not to Gaussian
-#' models. `"link"` will leave the values on scale of the linear predictors.
-#' `"response"` (or `NULL`) will transform them on scale of the response
-#' variable. Thus for a logistic model, `"link"` will give estimations expressed
-#' in log-odds (probabilities on logit scale) and `"response"` in terms of
-#' probabilities.
+#' See also section _Predictions on different scales_.
 #'
-#' To predict distributional parameters (called "dpar" in other packages), for
-#' instance when using complex formulae in `brms` models, the `predict` argument
-#' can take the value of the parameter you want to estimate, for instance
-#' `"sigma"`, `"kappa"`, etc.
-#'
-#' `"response"` and `"inverse_link"` both return predictions on the response
-#' scale, however, `"response"` first calculates predictions on the response
-#' scale for each observation and *then* aggregates them by groups or levels
-#' defined in `by`. `"inverse_link"` first calculates predictions on the link
-#' scale for each observation, then aggregates them by groups or levels defined
-#' in `by`, and finally back-transforms the predictions to the response scale.
-#' Both approaches have advantages and disadvantages. `"response"` usually
-#' produces less biased predictions, but confidence intervals might be outside
-#' reasonable bounds (i.e., for instance can be negative for count data). The
-#' `"inverse_link"` approach is more robust in terms of confidence intervals, but
-#' might produce biased predictions. In particular for mixed models, using
-#' `"response"` is recommended, because averaging across random effects groups
-#' is then more accurate.
 #' @param estimate The `estimate` argument determines how predictions are
 #' averaged ("marginalized") over variables not specified in `by` or `contrast`
 #' (non-focal predictors). It controls whether predictions represent a "typical"
@@ -144,12 +121,72 @@
 #' - Bayesian models: For Bayesian models, parameters are cleaned using
 #'   `describe_posterior()`, thus, arguments like, for example, `centrality`,
 #'   `rope_range`, or `test` are passed to that function.
+#' - Especially for `estimate_contrasts()` with integer focal predictors, for
+#'   which contrasts should be calculated, use argument `integer_as_numeric` to
+#'   set the maximum number of unique values in an integer predictor to treat
+#'   that predictor as "discrete inetger" or as numeric. For the first case,
+#'   contasts are calculated between values of the predictor, for the latter,
+#'   contrasts of slopes are calculated. If the integer has more than
+#'   `integer_as_numeric` unique values, it is treated as numeric. Defaults to
+#'   `5`.
 #'
 #' @inheritParams parameters::model_parameters.default
 #' @inheritParams estimate_expectation
-#' @inherit estimate_slopes details
 #'
-#' @return A data frame of estimated marginal means.
+#' @details
+#' The [estimate_slopes()], [estimate_means()] and [estimate_contrasts()]
+#' functions are forming a group, as they are all based on *marginal*
+#' estimations (estimations based on a model). All three are built on the
+#' **emmeans** or **marginaleffects** package (depending on the `backend`
+#' argument), so reading its documentation (for instance [emmeans::emmeans()],
+#' [emmeans::emtrends()] or this [website](https://marginaleffects.com/)) is
+#' recommended to understand the idea behind these types of procedures.
+#'
+#' - Model-based **predictions** is the basis for all that follows. Indeed,
+#' the first thing to understand is how models can be used to make predictions
+#' (see [estimate_link()]). This corresponds to the predicted response (or
+#' "outcome variable") given specific predictor values of the predictors (i.e.,
[…]
