+++ man/get_emmeans.Rd
@@ -17,7 +17,6 @@ get_emcontrasts(
   by = NULL,
   predict = NULL,
   comparison = "pairwise",
-  transform = NULL,
   keep_iterations = FALSE,
   verbose = TRUE,
   ...
@@ -27,7 +26,6 @@ get_emmeans(
   model,
   by = "auto",
   predict = NULL,
-  transform = NULL,
   keep_iterations = FALSE,
   verbose = TRUE,
   ...
@@ -120,31 +118,7 @@ transformation is selected (which usually is \code{"response"}). See also
 \href{https://CRAN.R-project.org/package=emmeans/vignettes/transformations.html}{this vignette}.
 }
 
-The \code{"link"} option does not apply to all models, and usually not to Gaussian
-models. \code{"link"} will leave the values on scale of the linear predictors.
-\code{"response"} (or \code{NULL}) will transform them on scale of the response
-variable. Thus for a logistic model, \code{"link"} will give estimations expressed
-in log-odds (probabilities on logit scale) and \code{"response"} in terms of
-probabilities.
-
-To predict distributional parameters (called "dpar" in other packages), for
-instance when using complex formulae in \code{brms} models, the \code{predict} argument
-can take the value of the parameter you want to estimate, for instance
-\code{"sigma"}, \code{"kappa"}, etc.
-
-\code{"response"} and \code{"inverse_link"} both return predictions on the response
-scale, however, \code{"response"} first calculates predictions on the response
-scale for each observation and \emph{then} aggregates them by groups or levels
-defined in \code{by}. \code{"inverse_link"} first calculates predictions on the link
-scale for each observation, then aggregates them by groups or levels defined
-in \code{by}, and finally back-transforms the predictions to the response scale.
-Both approaches have advantages and disadvantages. \code{"response"} usually
-produces less biased predictions, but confidence intervals might be outside
-reasonable bounds (i.e., for instance can be negative for count data). The
-\code{"inverse_link"} approach is more robust in terms of confidence intervals, but
-might produce biased predictions. In particular for mixed models, using
-\code{"response"} is recommended, because averaging across random effects groups
-is then more accurate.}
+See also section \emph{Predictions on different scales}.}
 
 \item{comparison}{Specify the type of contrasts or tests that should be
 carried out.
@@ -174,15 +148,6 @@ can be carried out within subsets by indicating the grouping variable
 after a vertical bar ( \code{|}).
 }
 }}
-
-\item{transform}{A function applied to predictions and confidence intervals
-to (back-) transform results, which can be useful in case the regression
-model has a transformed response variable (e.g., \code{lm(log(y) ~ x)}). For
-Bayesian models, this function is applied to individual draws from the
-posterior distribution, before computing summaries. Can also be \code{TRUE}, in
-which case \code{insight::get_transformation()} is called to determine the
-appropriate transformation-function. Note that no standard errors are returned
-when transformations are applied.}
 
 \item{keep_iterations}{If \code{TRUE}, will keep all iterations (draws) of
 bootstrapped or Bayesian models. They will be added as additional columns
@@ -215,6 +180,14 @@ Additional arguments can be passed to these functions.
 \item Bayesian models: For Bayesian models, parameters are cleaned using
 \code{describe_posterior()}, thus, arguments like, for example, \code{centrality},
 \code{rope_range}, or \code{test} are passed to that function.
+\item Especially for \code{estimate_contrasts()} with integer focal predictors, for
+which contrasts should be calculated, use argument \code{integer_as_numeric} to
+set the maximum number of unique values in an integer predictor to treat
+that predictor as "discrete inetger" or as numeric. For the first case,
[…]

[…]
