+++ R/estimate_means.R
@@ -124,8 +124,8 @@
 #' - Especially for `estimate_contrasts()` with integer focal predictors, for
 #'   which contrasts should be calculated, use argument `integer_as_numeric` to
 #'   set the maximum number of unique values in an integer predictor to treat
-#'   that predictor as "discrete inetger" or as numeric. For the first case,
-#'   contasts are calculated between values of the predictor, for the latter,
+#'   that predictor as "discrete integer" or as numeric. For the first case,
+#'   contrasts are calculated between values of the predictor, for the latter,
 #'   contrasts of slopes are calculated. If the integer has more than
 #'   `integer_as_numeric` unique values, it is treated as numeric. Defaults to
 #'   `5`.
