### 2025-03-30 7b8f06e
Revise docs, remove deprecated arg (#455)
* Revise docs, remove deprecated arg

* typo
