### 2025-03-30 6371c6c
Deploying to gh-pages from @ easystats/modelbased@7b8f06eb7a77067a850781db027d93d8bc581dff 🚀
