+++ DESCRIPTION
@@ -39,10 +39,10 @@ BugReports: https://github.com/easystats/modelbased/issues
 Depends:
     R (>= 3.6)
 Imports:
-    bayestestR (>= 0.15.1),
-    datawizard (>= 1.0.0),
-    insight (>= 1.0.1),
-    parameters (>= 0.24.1),
+    bayestestR (>= 0.15.2),
+    datawizard (>= 1.0.2),
+    insight (>= 1.1.0),
+    parameters (>= 0.24.2),
     graphics,
     stats,
     tools,
