### 2025-04-01 ea3321f
Draft vignette to show different behaviour for mixed models
