+++ vignettes/overview_of_vignettes.Rmd
@@ -36,6 +36,7 @@ All package vignettes are available at [https://easystats.github.io/modelbased/]
 * [What are, why use and how to get marginal means](https://easystats.github.io/modelbased/articles/estimate_means.html)
 * [Contrast analysis](https://easystats.github.io/modelbased/articles/estimate_contrasts.html)
 * [Marginal effects and derivatives](https://easystats.github.io/modelbased/articles/estimate_slopes.html)
+* [Mixed effects models](https://easystats.github.io/modelbased/articles/mixed_models.html)
 
 ### Interpretation
 
