### 2025-04-01 #7894 @spholmes
👋  @tjmahr : could you help us review this paper on the modelbased package?
Many thanks
Susan
