### 2025-04-01 #7894 @spholmes
@vincentarelbundock : the review is now active in the thread: https://github.com/openjournals/joss-reviews/issues/7969
if you don't see the checklist, please generate it with "@editorialbot  generate my checklist"...many thanks for your help.
