### 2025-04-01 d2bcfdb
remove redundant example
