### 2025-04-01 d8f7a8a
Draft vignette to show different behaviour for mixed models (#457)
* Draft vignette to show different behaviour for mixed models

* intro

* vig

* text

* vig

* ...

* remove redundant example

* ...

* finalize

* styler, wordlist

* update snapshots
