+++ vignettes/bibliography.bib
@@ -129,3 +129,18 @@
 	month = jan,
 	year = {2013}
 }
+
+
+@misc{heiss_marginal_2022,
+	title = {Marginal and conditional effects for {GLMMs} with \{marginaleffects\}},
+	copyright = {https://creativecommons.org/licenses/by/4.0/legalcode},
+	url = {https://www.andrewheiss.com/blog/2022/11/29/conditional-marginal-marginaleffects},
+	abstract = {As a field, statistics is really bad at naming things. Take, for instance, the term “fixed effects.” In econometrics and other social science-flavored statistics, this typically refers to categorical terms in a regression model.},
+	language = {en},
+	urldate = {2025-04-01},
+	journal = {Andrew Heiss's blog},
+	author = {Heiss, Andrew},
+	month = nov,
+	year = {2022},
+	doi = {10.59350/xwnfm-x1827},
+}
