+++ pkgdown/_pkgdown.yml
@@ -52,6 +52,8 @@ navbar:
           href: articles/estimate_contrasts.html
         - text: "Marginal effects and derivatives"
           href: articles/estimate_slopes.html
+        - text: "Mixed effects models"
+          href: articles/mixed_models.html
         - text: -------
         - text: "Interpretation"
         - text: "Use a model to make predictions"
