### 2025-04-01 4f47d13
Deploying to gh-pages from @ easystats/modelbased@d8f7a8a7dfbacf3c4aa4ec2aff773ce9773ccb23 🚀
