### 2025-04-01 #7969 @vincentarelbundock
@spholmes

FYI, I have completed the checklist.

I should disclose that I have collaborated with the authors on other open source projects in the past. This took the form of standard pull requests.

I do not have a personal or financial relationship with them, and I do not believe that there is an inherent or apparent conflict of interest.
