+++ vignettes/mixed_models.Rmd
@@ -50,14 +50,14 @@ To use the summaries from Heiss (2022):
 
 > This conditional vs. marginal distinction applies to any sort of hierarchical structure in multilevel models:
 >
-> - **Conditional effect** = the effect of a variable in an *average cluster* (i.e. group-specific, subject-specific or cluster-specific effect, or an average or a typical cluster)
-> - **Marginal effect** = effect of a variable *across clusters on average* (i.e. global/population-level effect, or clusters on average).
+> - **Conditional effect** = the effect of a variable in an *average cluster* (i.e., group-specific, subject-specific or cluster-specific effect, or an average or a typical cluster)
+> - **Marginal effect** = effect of a variable *across clusters on average* (i.e., global/population-level effect, or clusters on average).
 
 When working with mixed models, the `modelbased` package offers flexibility in calculating EMMs through different backends. The `"marginaleffects"` backend and the `"emmeans"` backend employ different underlying methodologies. **marginaleffects** typically focuses on marginal predictions by averaging over the random effects, while **emmeans** provides conditional predictions. Consequently, the EMMs obtained using these two backends may differ. As we will see, conditional and marginal predictions...
 
 - ... are usually similar for linear mixed models
 - ... are usually different for generalized linear mixed models
-- ... are usually different for both linear and generalized linear models, when the data is "imbalanced" and "average" predictions are requested (i.e. `estimate = "average"`); by _imbalanced_ we mean not equally distributed levels.
+- ... are usually different for both linear and generalized linear models, when the data is "imbalanced" and "average" predictions are requested (i.e., `estimate = "average"`); by _imbalanced_ we mean not equally distributed levels.
 
 In essence, the choice of backend and the understanding of whether we are looking at conditional or marginal predictions are critical for correctly interpreting the results of mixed models. Carefully considering the research question and the nature of the random effects will guide the selection of the appropriate approach.
 
@@ -94,7 +94,7 @@ estimate_means(model, "Days")
 estimate_means(model, "Days", backend = "emmeans")
 ```
 
-Here, we calculate the marginal predictions for the `Days` variable using the `estimate = "average"` argument. In this case, because the model is linear and the data is balanced, the predictions are again the same as the previous results, i.e. averaging across all observations makes no difference here.
+Here, we calculate the marginal predictions for the `Days` variable using the `estimate = "average"` argument. In this case, because the model is linear and the data is balanced, the predictions are again the same as the previous results, i.e., averaging across all observations makes no difference here.
 
 ```{r}
 # marginal predictions, averaged across all observations,
@@ -162,5 +162,11 @@ estimate_means(m1, "camper", estimate = "average", re.form = NA)
 ```
 
 This highlights that in GLMMs, the choice between marginal and conditional predictions significantly impacts the estimated means and, consequently, the interpretation of the model's results. In this case, the marginal predictions are higher than the conditional predictions.
+
+## Conclusion
+
+This vignette has demonstrated the nuances of calculating estimated marginal means (EMMs) in mixed models, highlighting the critical distinction between conditional and marginal predictions. While conditional predictions focus on specific levels of random effects, marginal predictions average over them, providing a population-level perspective. We've seen that in linear mixed models with balanced data, these two approaches often yield similar results. However, in generalized linear mixed models (GLMMs) and in linear mixed models with imbalanced data, the differences between conditional and marginal predictions can be substantial.
+
[…]
