### 2025-04-01 ca9304c
Deploying to gh-pages from @ easystats/modelbased@fcd2f7f49d4240f34f84023727f5716277b507a5 🚀
