+++ vignettes/mixed_models.Rmd
@@ -135,30 +135,30 @@ The generalized linear mixed model (GLMM) example, using a Poisson distribution 
 
 ```{r}
 data("fish", package = "insight")
-m1 <- lme4::glmer(
+model <- lme4::glmer(
   count ~ child + camper + (1 | persons),
   data = fish,
   family = poisson()
 )
 
 # marginal predictions
-estimate_means(m1, "camper")
+estimate_means(model, "camper")
 
 # conditional predictions
-estimate_means(m1, "camper", re.form = NA)
+estimate_means(model, "camper", re.form = NA)
 
 # conditional predictions
-estimate_means(m1, "camper", backend = "emmeans")
+estimate_means(model, "camper", backend = "emmeans")
 ```
 
 Furthermore, when using `estimate = "average"` to average across all observations, the differences between marginal and conditional predictions persist, with marginal predictions showing higher means (1.52 and 4.54) than conditional predictions (1.20 and 3.19), but both results differ from the predictions for "typical" observations based on the default `estimate` option (or `"emmeans"` backend), as shown in the previous output.
 
 ```{r}
 # average marginal predictions
-estimate_means(m1, "camper", estimate = "average")
+estimate_means(model, "camper", estimate = "average")
 
 # average conditional predictions
-estimate_means(m1, "camper", estimate = "average", re.form = NA)
+estimate_means(model, "camper", estimate = "average", re.form = NA)
 ```
 
 This highlights that in GLMMs, the choice between marginal and conditional predictions significantly impacts the estimated means and, consequently, the interpretation of the model's results. In this case, the marginal predictions are higher than the conditional predictions.
