### 2025-04-01 e293c1a
Deploying to gh-pages from @ easystats/modelbased@c36a9e86ad53f30168554cc21f51ffecd268b76e 🚀
