### 2025-04-01 2799a59
Deploying to gh-pages from @ easystats/modelbased@ad85a0d07a8336fdeecf8973bf02c14a40631140 🚀
