### 2025-04-01 #7894 @tjmahr
> 👋 [@tjmahr](https://github.com/tjmahr) : could you help us review this paper on the modelbased package? Many thanks Susan

Sure.
