### 2025-04-01 #7894 @bbolker
Hmm, I thought that you didn't need me because @vincentarelbundock jumped in. Sorry for confusion! If you still want another reviewer, please let me know ...
