### 2025-04-01 #7894 @spholmes
Sorry @bbolker, I thought it was clear we need two reviews for all JOSS papers and I invited each of you, I'll go with TJ for the time being and you can be our understudy...
thanks anyway
