### 2025-04-01 #7894 @spholmes
Thanks @tjmahr, the review is happening on the thread  https://github.com/openjournals/joss-reviews/issues/7969 where you should be able to ask the editorial bot
to generate your checklist with: " @editorialbot generate my checklist"...susan
