### 2025-04-02 #7894 @tjmahr
> Thanks [@tjmahr](https://github.com/tjmahr), the review is happening on the thread [#7969](https://github.com/openjournals/joss-reviews/issues/7969) where you should be able to ask the editorial bot to generate your checklist with: " [@editorialbot](https://github.com/editorialbot) generate my checklist"...susan

That didn't work.
