### 2025-04-03 #7969 @spholmes
Thanks @vincentarelbundock for your rapid review.
