### 2025-04-03 2523d8d
Trend within range
