+++ R/get_marginaltrends.R
@@ -24,16 +24,22 @@ get_marginaltrends <- function(model,
   model_info <- insight::model_info(model, verbose = FALSE)
 
   # Guess arguments
-  trend <- .guess_marginaltrends_arguments(model, trend, verbose, ...)
+  myargs <- .guess_marginaltrends_arguments(model, trend, by, verbose, ...)
 
   # First step: create a data grid --------------------------------------------
   # ---------------------------------------------------------------------------
 
   # data grid only when we have by predictors
-  if (is.null(by)) {
+  if (is.null(by) && is.null(myargs$range)) {
     datagrid <- datagrid_info <- NULL
   } else {
-    dg_args <- list(model, by = by, factors = "all", include_random = TRUE, verbose = FALSE)
+    dg_args <- list(
+      model,
+      by = c(by, myargs$range),
+      factors = "all",
+      include_random = TRUE,
+      verbose = FALSE
+    )
     # add user-arguments from "...", but remove those arguments that are already set
     dots[c("by", "factors", "include_random", "verbose")] <- NULL
     dg_args <- insight::compact_list(c(dg_args, dots))
@@ -41,6 +47,23 @@ get_marginaltrends <- function(model,
     # Get corresponding datagrid (and deal with particular ats)
     datagrid <- do.call(insight::get_datagrid, dg_args)
     datagrid_info <- attributes(datagrid)
+
+    # get "by" argument from data grid
+    myargs$by <- datagrid_info$at_specs$varname
+
+    # remove `range` (or: `trend`) variable from `by`. If user specified a range
+    # for `trend`, we need the data (the defined range) in the data grid, but we
+    # do not want to stratify by the same variable in `by` then (this won't
+    # work). instead, in `.guess_marginaltrends_arguments()`, we copy the
+    # `trend` value to `range`, which we pass to the data grid, and `trend` only
+    # contains the variable name after that. This variable needs to be removed,
+    # to not to be passed to `by`
+    if (!is.null(myargs$range)) {
+      myargs$by <- setdiff(myargs$by, myargs$trend)
+      if (!length(myargs$by)) {
+        myargs$by <- NULL
+      }
+    }
   }
 
   # Second step: prepare arguments for marginaleffects ------------------------
@@ -64,8 +87,8 @@ get_marginaltrends <- function(model,
   fun_args <- insight::compact_list(c(
     list(
       model,
-      variables = trend,
-      by = datagrid_info$at_specs$varname,
+      variables = myargs$trend,
+      by = myargs$by,
       newdata = datagrid,
       conf_level = ci
     ),
@@ -105,7 +128,7 @@ get_marginaltrends <- function(model,
     info = c(
       datagrid_info,
       list(
-        trend = trend,
+        trend = myargs$trend,
         datagrid = datagrid,
         coef_name = "Slope",
         p_adjust = p_adjust,
@@ -133,6 +156,7 @@ get_marginaltrends <- function(model,
 #' @keywords internal
 .guess_marginaltrends_arguments <- function(model,
                                             trend = NULL,
+                                            by = NULL,
                                             verbose = TRUE,
                                             ...) {
   # Gather info
@@ -152,6 +176,8 @@ get_marginaltrends <- function(model,
       insight::format_alert(paste0("No numeric variable was specified for slope estimation. Selecting `trend = \"", trend, "\"`.")) # nolint
     }
   }
+
+  # check that we have only one predictor
   if (length(trend) > 1) {
     if (verbose) {
       insight::format_alert(paste0(
@@ -162,7 +188,32 @@ get_marginaltrends <- function(model,
[…]
