### 2025-04-03 603e616
Trend within range (#461)
* Trend within range

* add tests

* styler
