+++ R/utils.R
@@ -3,7 +3,7 @@
 .brms_aux_elements <- function() {
   c(
     "sigma", "mu", "nu", "shape", "beta", "phi", "hu", "ndt", "zoi", "coi",
-    "kappa", "bias", "bs", "zi", "alpha", "xi"
+    "kappa", "bias", "bs", "zi", "alpha", "xi", "delta", "k"
   )
 }
 
