### 2025-04-03 2cc6fdb
Deploying to gh-pages from @ easystats/modelbased@8eea7d44527c0c686cb461eaa9bce4d21671cf0e 🚀
