### 2025-04-03 9a7a24c
Better error message Fixes #459
