### 2025-04-03 3bbcc61
Better error message (#462)
* Better error message
Fixes #459

* fix
