+++ tests/testthat/test-estimate_contrasts.R
@@ -913,3 +913,13 @@ test_that("estimate_contrast, don't calculate slopes for integers", {
   out <- estimate_contrasts(m, "hp", by = "gear")
   expect_identical(dim(out), c(3L, 8L))
 })
+
+
+test_that("estimate_contrast, informative error when `by` and `contrast` are the same", {
+  data(iris)
+  m <- lm(Petal.Length ~ Species, data = iris)
+  expect_error(
+    estimate_contrasts(m, "Species = 'versicolor'", by = "Species = 'setosa'"),
+    regex = "You cannot"
+  )
+})
