### 2025-04-03 3d0830f
correct affiliation and add JOSS GHA workflow
cf. https://github.com/openjournals/openjournals-draft-action
