+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,3 @@
-@ -1,30 +0,0 @@
 # TODO: delete this file once the paper is published
 on:
   push:
