### 2025-04-03 459035c
Deploying to gh-pages from @ easystats/modelbased@3d0830f9537c7eeef94352c06268f37811d2001a 🚀
