### 2025-04-03 3922941
Deploying to gh-pages from @ easystats/modelbased@6da30ece876f4d6dbd3ace1cf3584f4e93386061 🚀
