### 2025-04-04 11559c6
Revert "simplify reshape_grouplevel"
This reverts commit bdc3867998b1f3d2c639d3824c2a25a406596c21.
