### 2025-04-04 b0a7841
Reapply "simplify reshape_grouplevel"
This reverts commit 11559c6532db022ac6e4e1ce1db8270b645ca550.
