+++ R/utils.R
@@ -5,7 +5,7 @@
     "sigma", "mu", "nu", "shape", "beta", "phi", "hu", "ndt", "zoi", "coi",
     "kappa", "bias", "bs", "zi", "alpha", "xi", "delta", "k"
   )
-  if (!is.null(model)) {
+  if (inherits(model, "brmsfit")) {
     out <- unique(c(out, insight::find_auxiliary(model)))
   }
   out
