### 2025-04-06 1d89e2b
Allow individual dpars (#466)
* Allow individual dpars

* desc

* fix

* fix

* fix
