### 2025-04-08 b95bd03
error message (#468)
* error message

* simpler

* msg

* test
