### 2025-04-08 a2509b5
Deploying to gh-pages from @ easystats/modelbased@b95bd0341f29794d24ff874696ff99e9d0217723 🚀
