+++ R/reshape_grouplevel.R
@@ -1,15 +1,25 @@
 #' @rdname estimate_grouplevel
 #' @param x The output of `estimate_grouplevel()`.
-#' @param indices A list containing the indices (i.e., which columns) to
-#'   extract (e.g., "Coefficient", "Median").
-#' @param group A list containing the random factor to select (e.g., "Participant") if
-#'   the model was `y ~ x + (1|Participant)`.
+#' @param indices A character vector containing the indices (i.e., which
+#'   columns) to extract (e.g., "Coefficient", "Median").
+#' @param group The name of the random factor to select as string value (e.g.,
+#'   `"Participant"`, if the model was `y ~ x + (1|Participant)`.
+#'
 #' @export
-reshape_grouplevel <- function(x, indices = "all", group = "all", ...) {
+reshape_grouplevel <- function(x, ...) {
   UseMethod("reshape_grouplevel")
 }
 
+#' @export
+reshape_grouplevel.default <- function(x, ...) {
+  insight::format_error(paste0(
+    "`reshape_grouplevel()` not implemented yet for objects of class `",
+    class(x)[1],
+    "`."
+  ))
+}
 
+#' @rdname estimate_grouplevel
 #' @export
 reshape_grouplevel.estimate_grouplevel <- function(x, indices = "all", group = NULL, ...) {
   # Find indices
@@ -18,22 +28,33 @@ reshape_grouplevel.estimate_grouplevel <- function(x, indices = "all", group = N
   }
 
   # Accommodate Bayesian
-  if ("Coefficient" %in% indices) indices <- c(indices, "Median", "Mean", "MAP")
-  if ("SE" %in% indices) indices <- c(indices, "SD", "MAD")
+  if ("Coefficient" %in% indices) {
+    indices <- c(indices, "Median", "Mean", "MAP")
+  }
+  if ("SE" %in% indices) {
+    indices <- c(indices, "SD", "MAD")
+  }
   indices <- names(x)[names(x) %in% unique(indices)]
 
   # Random parameters
-  if (all(is.null(group))) group <- unique(x$Group)
-  if (length (group) > 1) {
-    warning(paste0("Multiple group are present (", paste(group, collapse = ", "),
-                   "). Selecting the first (group = ", group[1], ")."))
+  if (is.null(group)) {
+    group <- unique(x$Group)
+  }
+
+  if (length(group) > 1) {
+    insight::format_alert(paste0(
+      "Multiple groups are present (", toString(group),
+      "). Selecting the first (", group[1], ")."
+    ))
     group <- group[1]
   }
   x <- x[x$Group %in% group, ]
 
   # Create a new column for the parameter name
   x$.param <- x$Parameter
-  if("Component" %in% names(x)) x$.param <- paste0(x$Component, "_", x$.param)
+  if("Component" %in% names(x)) {
+    x$.param <- paste0(x$Component, "_", x$.param)
+  }
   x$.param <- gsub("conditional_", "", x$.param)
 
   # Reshape
