+++ R/get_emcontrasts.R
@@ -83,7 +83,7 @@ get_emcontrasts <- function(model,
   out <- emmeans::contrast(estimated, by = emm_by, method = comparison, ...)
 
   # for Bayesian model, keep iterations
-  if (insight::model_info(model)$is_bayesian) {
+  if (insight::model_info(model, response = 1)$is_bayesian) {
     attr(out, "posterior_draws") <- insight::get_parameters(estimated)
   } else {
     keep_iterations <- FALSE
@@ -143,7 +143,7 @@ get_emcontrasts <- function(model,
 
 .format_emmeans_contrasts <- function(model, estimated, ci, p_adjust, ...) {
   predict <- attributes(estimated)$predict
-  m_info <- insight::model_info(model)
+  m_info <- insight::model_info(model, response = 1)
 
   # Summarize and clean
   if (m_info$is_bayesian) {
