+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.10.0.13
+Version: 0.10.0.14
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
