### 2025-04-08 4e35a8f
Fix miltivariate brms models (#470)
* Fix miltivariate brms models

* news, desc
