+++ R/get_marginalcontrasts.R
@@ -37,7 +37,7 @@ get_marginalcontrasts <- function(model,
   on_the_fly_factors <- attributes(model_data)$factors
 
   # model details
-  model_info <- insight::model_info(model, verbose = FALSE)
+  model_info <- insight::model_info(model, response = 1, verbose = FALSE)
 
   # Guess arguments
   my_args <- .guess_marginaleffects_arguments(
