### 2025-04-08 e5cb122
Deploying to gh-pages from @ easystats/modelbased@4e35a8fc1ca96473fa1bc8fdae63d14e252269f5 🚀
