+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.10.0.15
+Version: 0.10.0.16
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -28,7 +28,7 @@ Authors@R:
              role = "aut",
              email = "remi.theriault@mail.mcgill.ca",
              comment = c(ORCID = "0000-0003-4315-6788")))
-Maintainer: Dominique Makowski <dom.makowski@gmail.com>
+Maintainer: Dominique Makowski <officialeasystats@gmail.com>
 Description: Implements a general interface for model-based estimations
     for a wide variety of models, used in the computation of
     marginal means, contrast analysis and predictions. For a list of supported models,
