### 2025-04-09 efeafa5
Deploying to gh-pages from @ easystats/modelbased@f9e4df6a87f4daf46b8d75fd3720b85c73279482 🚀
