### 2025-04-09 50c1259
Support Wiener (#471)
* Support Wiener

* add test
