+++ tests/testthat/_snaps/brms.md
@@ -1,0 +1,56 @@
+# estimate_means - brms, Wiener
+
+    Code
+      print(out)
+    Output
+      Model-based Predictions: rt
+      
+      rt   | Predicted |   SE |       95% CI
+      --------------------------------------
+      0.96 |      1.17 | 0.84 | [0.33, 3.54]
+      0.69 |      1.17 | 0.82 | [0.32, 3.24]
+      1.18 |      1.11 | 0.77 | [0.33, 3.30]
+      1.54 |      1.18 | 0.86 | [0.31, 3.44]
+      1.48 |      1.18 | 0.80 | [0.32, 3.35]
+      
+      Model-based Predictions: response
+      
+      response | Predicted |   SE |       95% CI
+      ------------------------------------------
+      1.00     |      0.42 | 0.49 | [0.00, 1.00]
+      0.00     |      0.42 | 0.49 | [0.00, 1.00]
+      0.00     |      0.41 | 0.49 | [0.00, 1.00]
+      0.00     |      0.43 | 0.50 | [0.00, 1.00]
+      0.00     |      0.40 | 0.49 | [0.00, 1.00]
+      
+      Variable predicted: rt, response
+      Predictions are on the response-scale.
+
+---
+
+    Code
+      print(out, table_width = Inf)
+    Output
+      Model-based Predictions: rt
+      
+      rt   | Predicted |   SE |       95% CI | iter_1 | iter_2 | iter_3
+      -----------------------------------------------------------------
+      0.96 |      1.26 | 0.95 | [0.33, 3.80] |   2.67 |   1.87 |   0.34
+      0.69 |      1.12 | 0.81 | [0.31, 3.15] |   3.59 |   0.44 |   2.17
+      1.18 |      1.18 | 0.92 | [0.32, 3.71] |   2.15 |   0.79 |   0.41
+      1.54 |      1.20 | 0.90 | [0.30, 3.81] |   1.08 |   0.70 |   0.37
+      1.48 |      1.19 | 0.89 | [0.32, 3.62] |   0.59 |   1.31 |   2.65
+      
+      Model-based Predictions: response
+      
+      response | Predicted |   SE |       95% CI | iter_1 | iter_2 | iter_3
+      ---------------------------------------------------------------------
+      1.00     |      0.42 | 0.49 | [0.00, 1.00] |   1.00 |   1.00 |   0.00
+      0.00     |      0.40 | 0.49 | [0.00, 1.00] |   1.00 |   1.00 |   1.00
+      0.00     |      0.41 | 0.49 | [0.00, 1.00] |   1.00 |   0.00 |   0.00
+      0.00     |      0.40 | 0.49 | [0.00, 1.00] |   0.00 |   0.00 |   0.00
+      0.00     |      0.40 | 0.49 | [0.00, 1.00] |   0.00 |   0.00 |   0.00
+      
+      Variable predicted: rt, response
+      Predictions are on the response-scale.
+
