+++ NEWS.md
@@ -9,6 +9,9 @@
 * Added `pool_slopes()`, to pool results from `estimate_slopes()` applied to
   imputed data.
 
+* `estimate_prediction()` and `estimate_relation()` now support Wiener-models
+  (Drift Diffusion Models) from package *brms*.
+
 * Column names of predicted values for `backend = "emmeans"` has changed for
   models like logistic regression, or beta regression. Formerly, name was
   `Mean`, now it is `Probability` or `Proportion`, depending on the model.
