+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.10.0.16
+Version: 0.10.0.17
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -87,6 +87,7 @@ Suggests:
     rmarkdown,
     rstanarm,
     rtdists,
+    RWiener,
     sandwich,
     see (>= 0.11.0),
     testthat (>= 3.2.1),
