### 2025-04-09 241cc40
Deploying to gh-pages from @ easystats/modelbased@50c125994b9adf529d3ab12da5775d416b4f3f01 🚀
