### 2025-04-09 0b7c0cc
Deploying to gh-pages from @ easystats/modelbased@7f81ea86bdaad893ada0e4017ec904be583210f2 🚀
