### 2025-04-09 84802cc
Deploying to gh-pages from @ easystats/modelbased@4e417988bb30a41b5cc8e737770ddece32efe3b3 🚀
