### 2025-04-10 def6c62
Make `predict` work in `estimate_slopes()` Fixes #472
