### 2025-04-10 cd55154
Make `predict` work in `estimate_slopes()` (#473)
* Make `predict` work in `estimate_slopes()`
Fixes #472

* typo

* use seed
