+++ R/estimate_slopes.R
@@ -91,6 +91,7 @@
 estimate_slopes <- function(model,
                             trend = NULL,
                             by = NULL,
+                            predict = NULL,
                             ci = 0.95,
                             p_adjust = "none",
                             transform = NULL,
@@ -108,6 +109,7 @@ estimate_slopes <- function(model,
     estimated <- get_emtrends(
       model,
       trend = trend,
+      predict = predict,
       by = by,
       keep_iterations = keep_iterations,
       verbose = verbose,
@@ -120,6 +122,7 @@ estimate_slopes <- function(model,
       model,
       trend = trend,
       by = by,
+      predict = predict,
       ci = ci,
       p_adjust = p_adjust,
       transform = transform,
