### 2025-04-11 43e270c
fix weights
