### 2025-04-11 243a975
fix weights (#475)
