### 2025-04-11 b882b50
Deploying to gh-pages from @ easystats/modelbased@243a9754916eaa92039503598cda86c600afa580 🚀
