### 2025-04-13 3a3ec0b
estimate_prediction() is extremely slow compared to insight::get_predicted() for Wiener Fixes #476
