+++ R/estimate_predicted.R
@@ -179,6 +179,12 @@
 #' be `TRUE`, in which case `insight::get_transformation()` is called to
 #' determine the appropriate transformation-function. Note that no standard
 #' errors are returned when transformations are applied.
+#' @param iterations For Bayesian models, this corresponds to the number of
+#' posterior draws. If `NULL`, will use all the draws (one for each iteration of
+#' the model). For frequentist models, if not `NULL`, will generate bootstrapped
+#' draws, from which bootstrapped CIs will be computed. Use `keep_iterations` to
+#' control if and how many draws will be included in the return output (data
+#' frame), which can be used, for instance, for plotting.
 #' @param ... You can add all the additional control arguments from
 #' [insight::get_datagrid()] (used when `data = "grid"`) and
 #' [insight::get_predicted()].
@@ -235,6 +241,7 @@ estimate_expectation <- function(model,
                                  predict = "expectation",
                                  ci = 0.95,
                                  transform = NULL,
+                                 iterations = NULL,
                                  keep_iterations = FALSE,
                                  ...) {
   .estimate_predicted(
@@ -242,6 +249,7 @@ estimate_expectation <- function(model,
     data = data,
     by = by,
     ci = ci,
+    iterations = iterations,
     keep_iterations = keep_iterations,
     predict = predict,
     transform = transform,
@@ -258,6 +266,7 @@ estimate_link <- function(model,
                           predict = "link",
                           ci = 0.95,
                           transform = NULL,
+                          iterations = NULL,
                           keep_iterations = FALSE,
                           ...) {
   # reset to NULL if only "by" was specified
@@ -270,6 +279,7 @@ estimate_link <- function(model,
     data = data,
     by = by,
     ci = ci,
+    iterations = iterations,
     keep_iterations = keep_iterations,
     predict = predict,
     transform = transform,
@@ -285,6 +295,7 @@ estimate_prediction <- function(model,
                                 predict = "prediction",
                                 ci = 0.95,
                                 transform = NULL,
+                                iterations = NULL,
                                 keep_iterations = FALSE,
                                 ...) {
   .estimate_predicted(
@@ -292,6 +303,7 @@ estimate_prediction <- function(model,
     data = data,
     by = by,
     ci = ci,
+    iterations = iterations,
     keep_iterations = keep_iterations,
     predict = predict,
     transform = transform,
@@ -307,6 +319,7 @@ estimate_relation <- function(model,
                               predict = "expectation",
                               ci = 0.95,
                               transform = NULL,
+                              iterations = NULL,
                               keep_iterations = FALSE,
                               ...) {
   # reset to NULL if only "by" was specified
@@ -319,6 +332,7 @@ estimate_relation <- function(model,
     data = data,
     by = by,
     ci = ci,
+    iterations = iterations,
     keep_iterations = keep_iterations,
     predict = predict,
     transform = transform,
@@ -336,6 +350,7 @@ estimate_relation <- function(model,
                                 predict = "expectation",
                                 ci = 0.95,
                                 transform = NULL,
+                                iterations = NULL,
                                 keep_iterations = FALSE,
                                 ...) {
   # only "by" or "data", but not both
@@ -435,17 +450,13 @@ estimate_relation <- function(model,
     model,
     data = data,
     predict = predict,
-    ci = ci
+    ci = ci,
+    iterations = iterations
   )
   # for predicting grouplevel random effects, add "allow.new.levels"
   if (!is.null(grouplevel_effects) && any(grouplevel_effects %in% grid_specs$at_spec$varname)) {
     prediction_args$allow.new.levels <- TRUE
     dots$allow.new.levels <- NULL
-  }
[…]
