+++ R/estimate_predicted.R
@@ -150,7 +150,6 @@
 #' various examples, tutorials and usecases.
 #'
 #' @inheritParams get_emmeans
-#' @inheritParams bayestestR::describe_posterior
 #' @param data A data frame with model's predictors to estimate the response. If
 #' `NULL`, the model's data is used. If `"grid"`, the model matrix is obtained
 #' (through [insight::get_datagrid()]).
@@ -183,7 +182,7 @@
 #' posterior draws. If `NULL`, will use all the draws (one for each iteration of
 #' the model). For frequentist models, if not `NULL`, will generate bootstrapped
 #' draws, from which bootstrapped CIs will be computed. Use `keep_iterations` to
-#' control if and how many draws will be included in the return output (data
+#' control if and how many draws will be included in the returned output (data
 #' frame), which can be used, for instance, for plotting.
 #' @param ... You can add all the additional control arguments from
 #' [insight::get_datagrid()] (used when `data = "grid"`) and
@@ -358,6 +357,11 @@ estimate_relation <- function(model,
     insight::format_error("You can only specify one of `by` or `data`, but not both.")
   }
 
+  # keep_iterations cannot be larger than interations
+  if (!is.null(keep_iterations) && !is.null(iterations) && is.numeric(keep_iterations) && is.numeric(iterations)) { # nolint
+    insight::format_error("`keep_iterations` cannot be larger than `iterations`.")
+  }
+
   # call "get_data()" only once...
   model_data <- insight::get_data(model, verbose = FALSE)
   is_model <- insight::is_model(model)
@@ -453,11 +457,13 @@ estimate_relation <- function(model,
     ci = ci,
     iterations = iterations
   )
+
   # for predicting grouplevel random effects, add "allow.new.levels"
   if (!is.null(grouplevel_effects) && any(grouplevel_effects %in% grid_specs$at_spec$varname)) {
     prediction_args$allow.new.levels <- TRUE
     dots$allow.new.levels <- NULL
   }
+
   # get predictions
   predictions <- do.call(insight::get_predicted, c(prediction_args, dots))
   out <- as.data.frame(predictions, keep_iterations = keep_iterations)
@@ -496,6 +502,7 @@ estimate_relation <- function(model,
 
   # Store relevant information
   attr(out, "ci") <- ci
+  attr(out, "iterations") <- iterations
   attr(out, "keep_iterations") <- keep_iterations
   attr(out, "response") <- model_response
   attr(out, "transform") <- !is.null(transform)
