### 2025-04-14 18ad57f
estimate_prediction() is extremely slow compared to insight::get_predicted() for Wiener (#477)
* estimate_prediction() is extremely slow compared to insight::get_predicted() for Wiener
Fixes #476

* news, desc

* update

* news

* add tests

* add test, add check
