+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.10.0.21
+Version: 0.10.0.22
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -50,6 +50,7 @@ Imports:
 Suggests:
     BH,
     betareg,
+    boot,
     bootES,
     brms,
     coda,
