### 2025-04-14 830d27c
Update paper.Rmd (#474)
