### 2025-04-14 097c72f
Remove confusing "averaged" footer line for `estimate = "average"`
