### 2025-04-14 b67291d
handle offset usage
