### 2025-04-14 ea507f8
Remove confusing "averaged" footer line for `estimate = "average"` (#478)
* Remove confusing "averaged" footer line for `estimate = "average"`

* add test for offset

* typo

* comment

* handle offset usage

* verbose
