### 2025-04-16 acccf1e
improve handling of offsets
