+++ R/get_marginalmeans.R
@@ -50,16 +50,14 @@ get_marginalmeans <- function(model,
     c("typical", "population", "specific", "average")
   )
 
-  # inform user about appropriate use of offset-terms
-  if (estimate %in% c("specific", "typical") && !is.null(insight::find_offset(model)) && verbose) { # nolint
-    insight::format_alert("Model contains an offset-term, which is set to its mean value. To fully account for the individual offset-values per observation, use `estimate = \"average\"`.")
-  }
-
   # model details
   model_info <- insight::model_info(model, response = 1, verbose = FALSE)
 
   # Guess arguments
   my_args <- .guess_marginaleffects_arguments(model, by, verbose = verbose, ...)
+
+  # inform user about appropriate use of offset-terms
+  .check_offset(model, estimate, offset = dots$offset, verbose = verbose)
 
   # find default response-type, and get information about back transformation
   predict_args <- .get_marginaleffects_type_argument(
@@ -109,6 +107,14 @@ get_marginalmeans <- function(model,
     datagrid <- do.call(insight::get_datagrid, dg_args)
     datagrid_info <- attributes(datagrid)
 
+    # handle offsets
+    if (!is.null(dots$offset)) {
+      model_offset <- insight::find_offset(model)
+      if (!is.null(model_offset)) {
+        datagrid[[model_offset]] <- dots$offset
+      }
+    }
+
     # restore data types -  if we have defined numbers in `by`, like
     # `by = "predictor = 5"`, and `predictor` was a factor, it is returned as
     # numeric in the data grid. Fix this here, else marginal effects will fail
@@ -122,7 +128,7 @@ get_marginalmeans <- function(model,
   # --------------------------------------------------------------------------
 
   # remove user-arguments from "..." that will be used when calling marginaleffects
-  dots[c("by", "conf_level", "type", "digits", "bias_correction", "sigma")] <- NULL
+  dots[c("by", "conf_level", "type", "digits", "bias_correction", "sigma", "offset")] <- NULL
 
   # model df - can be passed via `...`
   if (is.null(dots$df)) {
@@ -455,4 +461,29 @@ get_marginalmeans <- function(model,
   contrast <- validate_arg(contrast, "contrast")
 
   list(by = by, contrast = contrast)
+}
+
+
+.check_offset <- function(model, estimate, offset = NULL, verbose = TRUE) {
+  # check if model has an offset at all
+  if (!is.null(insight::find_offset(model)) && verbose) {
+    msg <- NULL
+    if (is.null(offset)) {
+      # if no offset argument was specified, tell user what this means
+      msg <- switch(estimate,
+        specific = ,
+        typical = "Model contains an offset-term, which is set to its mean value. If you want to average predictions over the distribution of the offset (if appropriate), use `estimate = \"average\"`. If you want to fix the offset to a specific value, for instance `1`, use `offset = 1`.",
+        "Model contains an offset-term and you average predictions over the distribution of that offset. If you want to fix the offset to a specific value, for instance `1`, use `offset = 1` and set `estimate = \"typical\"`."
+      )
+    } else {
+      # if offset was specified, and estimate averages over predictions, tell this
+      msg <- switch(estimate,
+        average = ,
+        population = paste0("For `estimate = \"", estimate, "\"`, predictions are averaged over the distribution of the offset and the `offset` argument is ignored. If you want to fix the offset to a specific value, for instance `1`, use `offset = 1` and set `estimate = \"typical\"`.")
+      )
+    }
+    if (!is.null(msg)) {
+      insight::format_alert(msg)
+    }
+  }
 }
