+++ man/estimate_means.Rd
@@ -154,6 +154,9 @@ contrasts are calculated between values of the predictor, for the latter,
 contrasts of slopes are calculated. If the integer has more than
 \code{integer_as_numeric} unique values, it is treated as numeric. Defaults to
 \code{5}.
+\item For count regression models that use an offset term, use \verb{offset = <value>}
+to fix the offset at a specific value. Or use \code{estimate = "average"}, to
+average predictions over the distribution of the offset (if appropriate).
 }}
 }
 \value{
