### 2025-04-16 b90ebdb
improve handling of offsets (#479)
* improve handling of offsets

* docs

* fixes
