### 2025-04-16 7e8a176
extract contrast from atributes
