### 2025-04-16 8ad81f2
Deploying to gh-pages from @ easystats/modelbased@eae316f489f4ffaeac923917a31e52077c63b7f8 🚀
