### 2025-04-16 86b9c9a
keep_iterations < iterations but errors Fixes #480
