### 2025-04-16 f57523d
keep_iterations < iterations but errors (#481)
Fixes #480
