### 2025-04-21 5c1ae88
modelbased should preserve the "Row" column from get_predicted() for categorical / MV models Fixes #482
