+++ R/estimate_predicted.R
@@ -479,11 +479,8 @@ estimate_relation <- function(model,
     out[[model_response]] <- NULL
   }
 
-  # clean-up: remove "Row" variable (from ordinal and alike) for data grids
-  if (is_grid || inherits(data, "datagrid")) {
-    out[["Row"]] <- NULL
-  } else if ("Row" %in% colnames(out)) {
-    # keep row-column for user-specific "data"
+  # keep row-column, but make sure it's integer
+  if ("Row" %in% colnames(out)) {
     out[["Row"]] <- insight::format_value(out[["Row"]], protect_integers = TRUE)
   }
 
