+++ tests/testthat/test-multivariate_response.R
@@ -20,5 +20,5 @@ test_that("multivariate response", {
   expect_identical(dim(out), c(64L, 5L))
 
   out <- estimate_expectation(fm1, by = "cyl")
-  expect_identical(dim(out), c(6L, 4L))
+  expect_identical(dim(out), c(6L, 5L))
 })
