+++ tests/testthat/test-attributes_estimatefun.R
@@ -112,7 +112,7 @@ test_that("attributes_means", {
   expect_named(
     attributes(estim),
     c(
-      "names", "row.names", "class", "ci", "keep_iterations", "response",
+      "names", "class", "row.names", "ci", "keep_iterations", "response",
       "transform", "model", "datagrid", "focal_terms", "preserve_range",
       "table_title", "coef_name", "model_info", "table_footer", "adjusted_for",
       "at_specs", "at", "by", "reference", "data"
