+++ tests/testthat/test-keep_iterations.R
@@ -253,7 +253,7 @@ test_that("estimate_slopes() - posterior draws, get_predicted", {
   expect_named(
     attributes(out),
     c(
-      "names", "row.names", "class", "ci", "keep_iterations", "response",
+      "names", "class", "row.names", "ci", "keep_iterations", "response",
       "transform", "model", "datagrid", "focal_terms", "preserve_range",
       "table_title", "coef_name", "model_info", "table_footer", "adjusted_for",
       "at_specs", "at", "by", "reference", "data"
