### 2025-04-21 4260002
modelbased should preserve the "Row" column from get_predicted() for categorical / MV models (#483)
* modelbased should preserve the "Row" column from get_predicted() for categorical / MV models
Fixes #482

* fix

* add test

* comment

* news

* fix

* fix test

* update test

* always keep row

* fix tests

* fix name

* update test
