### 2025-04-22 f45414c
Allow `offset` argument for `estimate_relation()` and alike
