+++ R/estimate_predicted.R
@@ -186,7 +186,8 @@
 #' frame), which can be used, for instance, for plotting.
 #' @param ... You can add all the additional control arguments from
 #' [insight::get_datagrid()] (used when `data = "grid"`) and
-#' [insight::get_predicted()].
+#' [insight::get_predicted()]. Furthermore, for count regression models that use
+#' an offset term, use `offset = <value>` to fix the offset at a specific value.
 #'
 #' @return A data frame of predicted values and uncertainty intervals, with
 #' class `"estimate_predicted"`. Methods for [`visualisation_recipe()`][visualisation_recipe.estimate_predicted]
@@ -376,12 +377,14 @@ estimate_relation <- function(model,
       insight::find_offset(model)
     )
     model_response <- insight::find_response(model)
+    model_offset <- insight::find_offset(model)
     is_nullmodel <- isTRUE(.safe(insight::is_nullmodel(model)))
     grouplevel_effects <- insight::find_random(model, flatten = TRUE, split_nested = TRUE)
   } else {
     # for stuff like data frame, no response and no null model
     variables <- colnames(model_data)
     model_response <- NULL
+    model_offset <- NULL
     is_nullmodel <- FALSE
     grouplevel_effects <- NULL
   }
@@ -438,6 +441,11 @@ estimate_relation <- function(model,
     response <- data[[model_response]]
   } else {
     response <- NULL
+  }
+
+  # handle offsets
+  if (!is.null(dots$offset) && !is.null(model_offset)) {
+    data[[model_offset]] <- dots$offset
   }
 
   # Keep only predictors (and response) --------
