### 2025-04-22 4a16d83
Allow `offset` argument for `estimate_relation()` and alike (#484)
* Allow `offset` argument for `estimate_relation()` and alike

* use CRAN datawizard
