### 2025-04-23 6f0da73
Simplify reshape_grouplevel() (#464)
* Reapply "simplify reshape_grouplevel"

This reverts commit 11559c6532db022ac6e4e1ce1db8270b645ca550.

* Update estimate_grouplevel.Rd

* Update reshape_grouplevel.R

* fix

* version

* version bump

* trigger tests

* simplify

* remove browser()

* styler

* tests

* fix example

* Update estimate_grouplevel.Rd

* desc, news

* fix vignette

* fix

---------

Co-authored-by: Daniel <mail@danielluedecke.de>
