### 2025-04-23 e448499
Deploying to gh-pages from @ easystats/modelbased@6f0da737ac73485791b18eba7ebde368d5e34c2c 🚀
