### 2025-04-24 easystats/modelbased#486 @tjmahr
marginal effects doesn't use t statistic
The p-values in `estimate_slopes()` from marginal effects for simple slopes differ from other approaches, likely because p is based on z values instead of t values. This will be confusing if users can't get consistent p-values between marginal effects and other approaches.

``` r
mtcars$f_cyl <- factor(mtcars$cyl)
m1 <- lm(mpg ~ factor(cyl) * wt + hp, mtcars)
m2 <- lm(mpg ~ f_cyl * wt + hp, mtcars)

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI |     t |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.05, -2.98] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.49,  2.00] | -1.21 |  0.225
#> 8     | -2.17 | 0.85 | [-3.83, -0.50] | -2.55 |  0.011
#>
#> Marginal effects estimated for wt
#> Type of slope was dY/dX
marginaleffects::slopes(m2, variables = "wt", by = "f_cyl")
#>
#>  f_cyl Estimate Std. Error     z Pr(>|z|)    S 2.5 % 97.5 %
#>      4    -5.52       1.29 -4.26   <0.001 15.6 -8.05  -2.98
#>      6    -3.24       2.68 -1.21   0.2254  2.1 -8.49   2.00
#>      8    -2.17       0.85 -2.55   0.0108  6.5 -3.83  -0.50
#>
#> Term: wt
#> Type:  response
#> Comparison: dY/dX

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", backend = "emmeans", verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI | t(25) |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.18, -2.85] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.76,  2.27] | -1.21 |  0.237
#> 8     | -2.17 | 0.85 | [-3.92, -0.42] | -2.55 |  0.017
#>
#> Marginal effects estimated for wt
interactions::sim_slopes(m2, "wt", "f_cyl")
#> Warning: Johnson-Neyman intervals are not available for factor predictors or
#> moderators.
#> SIMPLE SLOPES ANALYSIS
#>
#> Slope of wt when f_cyl = 4:
#>
#>    Est.   S.E.   t val.      p
#> ------- ------ -------- ------
#>   -5.52   1.29    -4.26   0.00
#>
#> Slope of wt when f_cyl = 6:
#>
#>    Est.   S.E.   t val.      p
#> ------- ------ -------- ------
#>   -3.24   2.68    -1.21   0.24
#>
#> Slope of wt when f_cyl = 8:
#>
#>    Est.   S.E.   t val.      p
#> ------- ------ -------- ------
#>   -2.17   0.85    -2.55   0.02
summary(lm(mpg ~ 0 + f_cyl + f_cyl:wt + hp, mtcars))
#>
#> Call:
#> lm(formula = mpg ~ 0 + f_cyl + f_cyl:wt + hp, data = mtcars)
#>
#> Residuals:
#>    Min     1Q Median     3Q    Max
#> -3.855 -1.381 -0.312  1.291  4.893
#>
#> Coefficients:
#>           Estimate Std. Error t value Pr(>|t|)
#> f_cyl4    41.11394    3.13741  13.104 1.06e-12 ***
#> f_cyl6    32.58364    8.62826   3.776 0.000878 ***
#> f_cyl8    28.42651    4.17506   6.809 3.89e-07 ***
#> hp        -0.02229    0.01146  -1.945 0.063057 .
#> f_cyl4:wt -5.51603    1.29382  -4.263 0.000251 ***
#> f_cyl6:wt -3.24487    2.67681  -1.212 0.236762
#> f_cyl8:wt -2.16607    0.85004  -2.548 0.017351 *
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#>
#> Residual standard error: 2.327 on 25 degrees of freedom
#> Multiple R-squared:  0.9904, Adjusted R-squared:  0.9877
#> F-statistic: 366.8 on 7 and 25 DF,  p-value: < 2.2e-16
```
