### 2025-04-24 easystats/modelbased#486 @strengejacke
Thanks for this comment! There are several issues going on here...

### First

We create a data grid, which sets the `trend` variable to its mean value, when no range is specified. This leads to minor differences:

``` r
mtcars$f_cyl <- factor(mtcars$cyl)
m1 <- lm(mpg ~ factor(cyl) * wt + hp, mtcars)
m2 <- lm(mpg ~ f_cyl * wt + hp, mtcars)

marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl") |>
  as.data.frame() |>
  datawizard::data_extract("statistic")
#> [1] -4.263037 -1.212267 -2.548209

nd <- insight::get_datagrid(m2, by = "f_cyl")
nd
#>   f_cyl      wt       hp
#> 1     4 3.21725 146.6875
#> 2     6 3.21725 146.6875
#> 3     8 3.21725 146.6875

marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl", newdata = nd) |>
  as.data.frame() |>
  datawizard::data_extract("statistic")
#> [1] -4.263568 -1.212081 -2.547856
```

### Second

Unlike for `estimate_means()`, we did not use the degrees of freedom from the model, thus, we indeed return the default (`z`), instead of a `t` statistic.

### Third

Even if we would fix the 2nd point, this still don't work, obviously due to a bug in _marginaleffects_, or because the df's are ignored intentionally for slopes:

``` r
mtcars$f_cyl <- factor(mtcars$cyl)
m1 <- lm(mpg ~ factor(cyl) * wt + hp, mtcars)
m2 <- lm(mpg ~ f_cyl * wt + hp, mtcars)

marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl") |>
  as.data.frame() |>
  datawizard::data_extract("statistic")
#> [1] -4.263037 -1.212267 -2.548209

marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl", df = 10) |>
  as.data.frame() |>
  datawizard::data_extract("statistic")
#> [1] -4.263037 -1.212267 -2.548209
```

Note that `?marginaleffects::avg_slopes` has a documented `df` argument, but it doesn't seem to affect the results?
