### 2025-04-24 easystats/modelbased#486 @tjmahr
Right, but the table

```
modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI |     t |      p
```

says `t` for the statistic. That's the problem here.
