### 2025-04-24 easystats/modelbased#486 @strengejacke
You're right, my answer is nonsense. `df` should affect p-values/ci, not the test-statistic.
