### 2025-04-24 62153bd
Update NEWS.md
