### 2025-04-24 easystats/modelbased#486 @strengejacke
This is the solution now, consistent with `estimate_means()` and `estimate_contrasts()`. By default, the model's "Wald" degrees of freedom are extracted (residual df for t-models, Inf for z-models). Else, specify the `df` argument:

``` r
mtcars$f_cyl <- factor(mtcars$cyl)
m2 <- lm(mpg ~ f_cyl * wt + hp, mtcars)

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI | t(25) |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.18, -2.85] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.76,  2.27] | -1.21 |  0.237
#> 8     | -2.17 | 0.85 | [-3.92, -0.42] | -2.55 |  0.017
#>
#> Marginal effects estimated for wt
#> Type of slope was dY/dX

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", df = Inf, verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI |     z |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.05, -2.98] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.49,  2.00] | -1.21 |  0.225
#> 8     | -2.17 | 0.85 | [-3.83, -0.50] | -2.55 |  0.011
#>
#> Marginal effects estimated for wt
#> Type of slope was dY/dX

modelbased::estimate_means(m2, c("wt=[sd]", "f_cyl"), verbose = TRUE)
#> Estimated Marginal Means
#>
#> wt   | f_cyl |  Mean |   SE |         95% CI | t(25)
#> ----------------------------------------------------
#> 2.24 | 4     | 25.49 | 1.02 | [23.39, 27.59] | 25.01
#> 3.22 | 4     | 20.10 | 1.55 | [16.91, 23.28] | 13.00
#> 4.20 | 4     | 14.70 | 2.64 | [ 9.27, 20.13] |  5.58
#> 2.24 | 6     | 22.05 | 2.50 | [16.90, 27.20] |  8.81
#> 3.22 | 6     | 18.87 | 0.97 | [16.88, 20.87] | 19.50
#> 4.20 | 6     | 15.70 | 3.06 | [ 9.41, 21.99] |  5.14
#> 2.24 | 8     | 20.31 | 1.76 | [16.68, 23.94] | 11.52
#> 3.22 | 8     | 18.19 | 1.15 | [15.82, 20.56] | 15.79
#> 4.20 | 8     | 16.07 | 0.97 | [14.08, 18.06] | 16.64
#>
#> Variable predicted: mpg
#> Predictors modulated: wt=[sd], f_cyl
#> Predictors averaged: hp (1.5e+02)

modelbased::estimate_means(m2, c("wt=[sd]", "f_cyl"), df = Inf, verbose = TRUE)
#> Estimated Marginal Means
#>
#> wt   | f_cyl |  Mean |   SE |         95% CI |     z
#> ----------------------------------------------------
#> 2.24 | 4     | 25.49 | 1.02 | [23.50, 27.49] | 25.01
#> 3.22 | 4     | 20.10 | 1.55 | [17.07, 23.13] | 13.00
#> 4.20 | 4     | 14.70 | 2.64 | [ 9.53, 19.87] |  5.58
#> 2.24 | 6     | 22.05 | 2.50 | [17.14, 26.95] |  8.81
#> 3.22 | 6     | 18.87 | 0.97 | [16.98, 20.77] | 19.50
#> 4.20 | 6     | 15.70 | 3.06 | [ 9.71, 21.69] |  5.14
#> 2.24 | 8     | 20.31 | 1.76 | [16.85, 23.76] | 11.52
#> 3.22 | 8     | 18.19 | 1.15 | [15.93, 20.45] | 15.79
#> 4.20 | 8     | 16.07 | 0.97 | [14.18, 17.96] | 16.64
#>
#> Variable predicted: mpg
#> Predictors modulated: wt=[sd], f_cyl
#> Predictors averaged: hp (1.5e+02)
```
