+++ tests/testthat/_snaps/windows/print.new.md
@@ -1,584 +1,0 @@
-# estimate_slopes - print summary
-
-    Code
-      estimate_slopes(model_lm, trend = "x", by = "x", backend = "emmeans")
-    Output
-      Estimated Marginal Effects
-      
-      x     | Slope |   SE |         95% CI | t(998) |      p
-      -------------------------------------------------------
-      -3.38 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      -2.64 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      -1.90 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      -1.16 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      -0.42 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      0.32  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      1.06  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      1.80  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      2.54  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      3.28  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      
-      Marginal effects estimated for x
-
----
-
-    Code
-      summary(deriv)
-    Message
-      There might be too few data to accurately determine intervals. Consider
-        setting `length = 100` (or larger) in your call to `estimate_slopes()`.
-    Output
-      Johnson-Neymann Intervals
-      
-      Start |  End | Direction | Confidence 
-      --------------------------------------
-      -3.38 | 3.28 | positive  | Significant
-      
-      Marginal effects estimated for x
-
----
-
-    Code
-      estimate_slopes(model_lm, trend = "x", by = "x", backend = "marginaleffects")
-    Output
-      Estimated Marginal Effects
-      
-      x     | Slope |   SE |         95% CI | t(998) |      p
-      -------------------------------------------------------
-      -3.38 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      -2.64 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      -1.90 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      -1.16 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      -0.42 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      0.32  | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      1.06  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      1.80  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      2.54  | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
-      3.28  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
-      
-      Marginal effects estimated for x
-      Type of slope was dY/dX
-
----
-
-    Code
-      summary(deriv2)
-    Message
-      There might be too few data to accurately determine intervals. Consider
-        setting `length = 100` (or larger) in your call to `estimate_slopes()`.
-    Output
-      Johnson-Neymann Intervals
-      
-      Start |  End | Direction | Confidence 
-      --------------------------------------
-      -3.38 | 3.28 | positive  | Significant
-      
-      Marginal effects estimated for x
-      Type of slope was dY/dX
-
-# estimate_slopes - print regular
-
[…]
