### 2025-04-25 easystats/modelbased#486 @strengejacke
Output is consistent with _marginaleffects_ now:

``` r
mtcars$f_cyl <- factor(mtcars$cyl)
m2 <- lm(mpg ~ f_cyl * wt + hp, mtcars)

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI | t(25) |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.18, -2.85] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.76,  2.27] | -1.21 |  0.237
#> 8     | -2.17 | 0.85 | [-3.92, -0.42] | -2.55 |  0.017
#>
#> Marginal effects estimated for wt
#> Type of slope was dY/dX
marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl", df = 25)
#>
#>  f_cyl Estimate Std. Error     t Pr(>|t|)    S 2.5 % 97.5 % Df
#>      4    -5.52       1.29 -4.26   <0.001 12.0 -8.18 -2.851 25
#>      6    -3.24       2.68 -1.21   0.2367  2.1 -8.76  2.268 25
#>      8    -2.17       0.85 -2.55   0.0174  5.8 -3.92 -0.415 25
#>
#> Term: wt
#> Type:  response
#> Comparison: dY/dX

modelbased::estimate_slopes(m2, trend = "wt", by = "f_cyl", df = Inf, verbose = TRUE)
#> Estimated Marginal Effects
#>
#> f_cyl | Slope |   SE |         95% CI |     z |      p
#> ------------------------------------------------------
#> 4     | -5.52 | 1.29 | [-8.05, -2.98] | -4.26 | < .001
#> 6     | -3.24 | 2.68 | [-8.49,  2.00] | -1.21 |  0.225
#> 8     | -2.17 | 0.85 | [-3.83, -0.50] | -2.55 |  0.011
#>
#> Marginal effects estimated for wt
#> Type of slope was dY/dX
marginaleffects::avg_slopes(m2, variables = "wt", by = "f_cyl")
#>
#>  f_cyl Estimate Std. Error     z Pr(>|z|)    S 2.5 % 97.5 %
#>      4    -5.52       1.29 -4.26   <0.001 15.6 -8.05  -2.98
#>      6    -3.24       2.68 -1.21   0.2254  2.1 -8.49   2.00
#>      8    -2.17       0.85 -2.55   0.0108  6.5 -3.83  -0.50
#>
#> Term: wt
#> Type:  response
#> Comparison: dY/dX
```
