### 2025-04-25 b2d76b1
marginal effects doesn't use t statistic (#487)
* marginal effects doesn't use t statistic
Fixes #486

* fix

* Update format.R

* Update NEWS.md

* update test

* fix tests

* update

* update snaps

* fix

* fix tests

* Update test-estimate_slopes.R

* fix
