+++ tests/testthat/_snaps/windows/print.md
@@ -43,18 +43,18 @@
     Output
       Estimated Marginal Effects
       
-      x     | Slope |   SE |         95% CI |     t |      p
-      ------------------------------------------------------
-      -3.38 | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      -2.64 | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      -1.90 | 12.75 | 0.25 | [12.26, 13.24] | 50.97 | < .001
-      -1.16 | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      -0.42 | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      0.32  | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      1.06  | 12.75 | 0.25 | [12.26, 13.24] | 50.97 | < .001
-      1.80  | 12.75 | 0.25 | [12.26, 13.24] | 50.97 | < .001
-      2.54  | 12.75 | 0.25 | [12.26, 13.24] | 50.98 | < .001
-      3.28  | 12.75 | 0.25 | [12.26, 13.24] | 50.97 | < .001
+      x     | Slope |   SE |         95% CI | t(998) |      p
+      -------------------------------------------------------
+      -3.38 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      -2.64 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      -1.90 | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
+      -1.16 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      -0.42 | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      0.32  | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      1.06  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
+      1.80  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
+      2.54  | 12.75 | 0.25 | [12.26, 13.24] |  50.98 | < .001
+      3.28  | 12.75 | 0.25 | [12.26, 13.24] |  50.97 | < .001
       
       Marginal effects estimated for x
       Type of slope was dY/dX
@@ -96,9 +96,9 @@
     Output
       Estimated Marginal Effects
       
-      Slope |   SE |       95% CI |    t |      p
-      -------------------------------------------
-      0.79  | 0.10 | [0.59, 0.99] | 7.69 | < .001
+      Slope |   SE |       95% CI | t(144) |      p
+      ---------------------------------------------
+      0.79  | 0.10 | [0.59, 0.99] |   7.69 | < .001
       
       Marginal effects estimated for Petal.Length
       Type of slope was dY/dX
