### 2025-04-25 bdb4ce0
estimate_contrast(): accept any "predict" + backends divergence Fixes #488
