### 2025-04-25 0dec086
Add reference to "The New Statistics manifesto"
