+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{cumming2014new,
+  title={The new statistics: Why and how},
+  author={Cumming, Geoff},
+  journal={Psychological science},
+  volume={25},
+  number={1},
+  pages={7--29},
+  year={2014},
+  publisher={Sage Publications Sage CA: Los Angeles, CA}
+}
+
 @article{OSC2015estimating,
   title ={Estimating the reproducibility of psychological science},
   author = {{Open Science Collaboration}},
