### 2025-04-25 374df0e
mention interface to backend and plotting capabilities
