### 2025-04-25 1a551a0
add shrinkage citation
