### 2025-04-27 010ee1f
estimate_contrast(): accept any "predict" + backends divergence (#489)
* estimate_contrast(): accept any "predict" + backends divergence
Fixes #488

* make dpars work

* fix

* Update clean_names.R

* Update get_marginalmeans.R

* Update get_marginalmeans.R

* uodate tests

* Update test-predict-dpar.R
