### 2025-04-27 d8204ad
Deploying to gh-pages from @ easystats/modelbased@010ee1f4b61eed5973a2f2a9b7fcc3a7321ca15e 🚀
