### 2025-04-28 6422d16
move technical details to README
