+++ vignettes/overview_of_vignettes.Rmd
@@ -50,10 +50,6 @@ All package vignettes are available at [https://easystats.github.io/modelbased/]
 * [Visualize effects and interactions](https://easystats.github.io/modelbased/articles/estimate_relation.html)
 * [The Modelisation Approach to Statistics](https://easystats.github.io/modelbased/articles/modelisation_approach.html)
 
-### Technical Details
-
-* [Technical details about the different backends](https://easystats.github.io/modelbased/articles/technical_details.html)
-
 
 ## Case Studies
 
