### 2025-04-28 a5a840c
JOSS reviewer 2 (#490)
* Add reference to "The New Statistics manifesto"

* mention interface to backend and plotting capabilities

* replace "withhold", remove "vast" and reverse previous change

* typo

* add shrinkage citation

* add vignette

* move technical details to README

* knit readme

* wordlist

* add comment

* wordlist

* check

---------

Co-authored-by: Daniel <mail@danielluedecke.de>
