+++ paper/paper.bib
@@ -1,3 +1,24 @@
+@article{cumming2014new,
+  title={The new statistics: Why and how},
+  author={Cumming, Geoff},
+  journal={Psychological science},
+  volume={25},
+  number={1},
+  pages={7--29},
+  year={2014},
+  publisher={Sage Publications Sage CA: Los Angeles, CA}
+}
+
+@article{pan2014random,
+  title={Random effects selection in generalized linear mixed models via shrinkage penalty function},
+  author={Pan, Jianxin and Huang, Chao},
+  journal={Statistics and Computing},
+  volume={24},
+  pages={725--738},
+  year={2014},
+  publisher={Springer}
+}
+
 @article{OSC2015estimating,
   title ={Estimating the reproducibility of psychological science},
   author = {{Open Science Collaboration}},
