+++ vignettes/overview_of_vignettes.Rmd
@@ -50,6 +50,7 @@ All package vignettes are available at [https://easystats.github.io/modelbased/]
 * [Visualize effects and interactions](https://easystats.github.io/modelbased/articles/estimate_relation.html)
 * [The Modelisation Approach to Statistics](https://easystats.github.io/modelbased/articles/modelisation_approach.html)
 
+
 ## Case Studies
 
 ### Workflows
