### 2025-04-28 155f616
Deploying to gh-pages from @ easystats/modelbased@a5a840cfe92e0720e8c7ed7bc67d3d85439b84a5 🚀
