### 2025-04-28 02a1b26
Deploying to gh-pages from @ easystats/modelbased@65326dce16d19ad6e7cd2c7e5587e4e0a0d37576 🚀
