+++ DESCRIPTION
@@ -39,10 +39,10 @@ BugReports: https://github.com/easystats/modelbased/issues
 Depends:
     R (>= 3.6)
 Imports:
-    bayestestR (>= 0.15.2),
+    bayestestR (>= 0.15.3),
     datawizard (>= 1.0.2),
-    insight (>= 1.1.0),
-    parameters (>= 0.24.2),
+    insight (>= 1.2.0),
+    parameters (>= 0.25.0),
     graphics,
     stats,
     tools,
