### 2025-05-02 #7969 @strengejacke
Hi @spholmes,
we received feedback from both reviewers, which we have addressed in these two PR:
https://github.com/easystats/modelbased/pull/474
https://github.com/easystats/modelbased/pull/490

Furthermore, one of the reviewers filed an issue (https://github.com/easystats/modelbased/issues/486), which we also fixed.

Finally, we strongly shortened the paper, moving a whole section into the readme, to address the comments regarding the length of the paper. All changes are merged into the main branch, so the paper.md file can be here.
