### 2025-05-02 c8cb503
prepare CRAN release (#492)
* prepare CRAN release

* update version requirements

* use remotes for now, for checks

* Update test-backtransform_invlink.R

* update snaps

* remove remotes
