### 2025-05-04 e62b1fc
Add tests and sample data (#494)
* Add tests and sample data

* document example data

* try to fix

* test
