+++ man/contrast_example.Rd
@@ -1,0 +1,10 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{contrast_example}
+\alias{contrast_example}
+\title{Sample dataset}
+\description{
+Just for internal testing
+}
+\keyword{data}
