### 2025-05-04 ba9b027
Deploying to gh-pages from @ easystats/modelbased@4c28abe8af58408b8bdfbda012820174e4df95f6 🚀
