+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.11.0
+Version: 0.11.0.1
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -105,3 +105,4 @@ Roxygen: list(markdown = TRUE)
 Config/Needs/check: stan-dev/cmdstanr
 Config/Needs/website: easystats/easystatstemplate
 LazyData: true
+Remotes: easystats/insight, easystats/bayestestR, easystats/parameters
