### 2025-05-05 ee583f6
Deploying to gh-pages from @ easystats/modelbased@0b48a256e142ec1df8a33b5309983ce986f06bbf 🚀
