### 2025-05-06 73eb5ae
`comparison` can be a function or matrix
