### 2025-05-06 0e47185
`comparison` can be a function or matrix (#498)
