+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.11.0.2
+Version: 0.11.0.3
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
