### 2025-05-09 f13ae99
Allow joint-test for factorial designs Fixes #500
