### 2025-05-10 8533a31
check imput, add test
