### 2025-05-10 fd0f22d
add emmeans method
