+++ tests/testthat/test-joint_test.R
@@ -16,24 +16,42 @@ test_that("estimate_contrasts - joint test, 2-way", {
     comparison = "joint"
   )
   out2 <- emmeans::joint_tests(m, "coffee")
+  out3 <- estimate_contrasts(
+    m,
+    contrast = "time",
+    by = "coffee",
+    comparison = "joint",
+    backend = "emmeans"
+  )
 
   expect_identical(out1$coffee, out2$coffee)
   expect_equal(out1$`F`, out2$`F.ratio`, tolerance = 1e-3)
   expect_equal(out1$p, out2$p.value, tolerance = 1e-3)
+  expect_equal(out3$Statistic, out2$`F.ratio`, tolerance = 1e-3)
+  expect_equal(out3$p, out2$p.value, tolerance = 1e-3)
   expect_named(out1, c("Contrast", "coffee", "df1", "df2", "Difference", "F", "p"))
+  expect_named(out3, c("Contrast", "coffee", "df1", "df2", "F", "p"))
 
   out1 <- estimate_contrasts(
     m,
     contrast = "coffee",
     by = "time",
-    comparison = "joint",
-    joint_test = "chisq"
+    comparison = "joint"
   )
   out2 <- emmeans::joint_tests(m, "time")
+  out3 <- estimate_contrasts(
+    m,
+    contrast = "coffee",
+    by = "time",
+    comparison = "joint",
+    backend = "emmeans"
+  )
 
   expect_identical(out1$time, out2$time)
   expect_equal(out1$`F`, out2$`F.ratio`, tolerance = 1e-3)
   expect_equal(out1$p, out2$p.value, tolerance = 1e-3)
+  expect_equal(out3$Statistic, out2$`F.ratio`, tolerance = 1e-3)
+  expect_equal(out3$p, out2$p.value, tolerance = 1e-3)
 })
 
 
@@ -93,6 +111,13 @@ test_that("estimate_contrasts - joint test, 3-way", {
   )
   out2 <- emmeans::joint_tests(m, c("coffee", "sex"))
   out2 <- as.data.frame(out2[c(1, 3, 2, 4), ])
+  out3 <- estimate_contrasts(
+    m,
+    contrast = "time",
+    by = c("coffee", "sex"),
+    comparison = "joint",
+    backend = "emmeans"
+  )
 
   expect_identical(out1$coffee, out2$coffee)
   expect_identical(out1$sex, out2$sex)
