+++ R/get_emcontrasts.R
@@ -98,7 +98,7 @@ get_emcontrasts <- function(model,
 
   # intermediate step: joint tests
   if (joint_test) {
-    out <- .joint_test(out)
+    out <- .joint_test(out, my_args)
   }
 
   # for Bayesian model, keep iterations
