### 2025-05-10 02befcd
capture error
