### 2025-05-10 6253c72
Allow joint-test for factorial designs (#501)
* Allow joint-test for factorial designs
Fixes #500

* fix

* fix

* ...

* update

* fix

* update

* fix

* fix

* finally?

* version

* pass test arg, add tests

* check imput, add test

* news

* add emmeans method

* add tests

* add test

* fix

* fix test

* fix test

* fix

* capture error

* msg

* fix tests

* minor

* code style
