### 2025-05-10 d5ba079
Deploying to gh-pages from @ easystats/modelbased@fa86c20079e80476cab5c5ed66f75db3f35a00ab 🚀
