### 2025-05-11 7b287d2
Add vignette on custom contrasts and joint tests
