+++ vignettes/introduction_comparisons_2.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Slopes, floodlight and spotlight analysis (Johnson-Neyman intervals)"
+title: "Contrasts and pairwise comparisons"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Slopes, floodlight and spotlight analysis (Johnson-Neyman intervals)}
+  %\VignetteIndexEntry{Contrasts and pairwise comparisons}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -19,12 +19,13 @@ knitr::opts_chunk$set(
   package.startup.message = FALSE
 )
 options(width = 800)
-options(modelbased_join_dots = FALSE)
-ht6 <- m <- NULL
-
 arrow_color <- "#FF00cc"
+my_predictions <- NULL
 
-pkgs <- c("ggplot2", "see", "marginaleffects", "parameters")
+pkgs <- c("ggplot2", "marginaleffects", "see", "parameters")
+
+options(modelbased_join_dots = FALSE)
+options(modelbased_select = "minimal")
 
 if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
   knitr::opts_chunk$set(eval = FALSE)
@@ -32,243 +33,326 @@ if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
 ```
 
 
-This vignette is the second in a 4-part series:
+This vignette is the second in a 5-part series:
 
 1. [**Contrasts and Pairwise Comparisons**](https://easystats.github.io/modelbased/articles/introduction_comparisons_1.html)
 
-2. **Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**
+2. **User Defined Contrasts and Joint Tests**
 
-3. [**Contrasts and Comparisons for Generalized Linear Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
+3. [**Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
 
-4. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_4.html)
+4. [**Contrasts and Comparisons for Generalized Linear Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_4.html)
 
-# Contrasts and comparisons for slopes of numeric predictors
+5. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_5.html)
 
-For numeric focal terms, it is possible to calculate contrasts for slopes, or the _linear trend_ of these focal terms. Let's start with a simple example again.
+
+data(contrast_example, package = "modelbased")
+
+  # set contrasts.
+  treat_vs_none <- c(-2 / 3, 1 / 3, 1 / 3)
+  short_vs_long <- c(0, -1 / 2, 1 / 2)
+
+  contrast_example$tx_ori <- contrast_example$tx
+  contrasts(contrast_example$tx) <- cbind(treat_vs_none, short_vs_long)
+
+  contrast_example$treat_vs_none <- as.factor(
+    ifelse(contrast_example$tx == "no treatment", "no treatment", "Puppies")
+  )
+
+  contrast_example$short_vs_long <- factor(
+    short_vs_long,
+    levels = c("short", "long", "no treatmen")
+  )
+
[…]
