+++ vignettes/introduction_comparisons_2.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Contrasts and pairwise comparisons"
+title: "User Defined Contrasts and Joint Tests"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Contrasts and pairwise comparisons}
+  %\VignetteIndexEntry{User Defined Contrasts and Joint Tests}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -22,7 +22,7 @@ options(width = 800)
 arrow_color <- "#FF00cc"
 my_predictions <- NULL
 
-pkgs <- c("ggplot2", "marginaleffects", "see", "parameters")
+pkgs <- c("marginaleffects", "parameters")
 
 options(modelbased_join_dots = FALSE)
 options(modelbased_select = "minimal")
@@ -46,310 +46,60 @@ This vignette is the second in a 5-part series:
 5. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_5.html)
 
 
-data(contrast_example, package = "modelbased")
+This vignette demonstrates how to define and apply user-defined contrasts to a factor variable within a linear model.
 
-  # set contrasts.
-  treat_vs_none <- c(-2 / 3, 1 / 3, 1 / 3)
-  short_vs_long <- c(0, -1 / 2, 1 / 2)
+First, we define two specific contrasts: `treat_vs_none` (comparing the average of two treatment levels to a control level) and `short_vs_long` (comparing two treatment levels against each other).
 
-  contrast_example$tx_ori <- contrast_example$tx
-  contrasts(contrast_example$tx) <- cbind(treat_vs_none, short_vs_long)
+To illustrate the effect of these custom contrasts, the code creates a copy of the original treatment factor (`tx_ori`) before applying the new contrasts to the tx factor. It also centers a continuous predictor, `score`, to better interpret the main effects of the interaction term.
 
-  contrast_example$treat_vs_none <- as.factor(
-    ifelse(contrast_example$tx == "no treatment", "no treatment", "Puppies")
-  )
+Two linear models are then fitted: `m1` uses the `tx` factor with the user-defined contrasts, while `m2` uses the original `tx_ori` factor. The `compare_parameters()` function is used to show that `m1` directly estimates coefficients corresponding to `treat_vs_none` and `short_vs_long`, which are not directly available in `m2`.
 
-  contrast_example$short_vs_long <- factor(
-    short_vs_long,
-    levels = c("short", "long", "no treatmen")
-  )
-
-  contrast_example$score <- contrast_example$score - mean(contrast_example$score)
-
-  # fit model
-  m1 <- lm(outcome ~ score + tx + tx:score, data = contrast_example)
-  m2 <- lm(outcome ~ score * tx_ori, data = contrast_example)
-
-  expect_equal(
-    coef(m1)["txtreat_vs_none"],
-    estimate_contrasts(m2, "tx_ori", comparison = "((b2+b3)/2) = b1")$Difference,
-    tolerance = 1e-4,
-    ignore_attr = TRUE
-  )
-  expect_equal(
-    coef(m1)["txshort_vs_long"],
-    estimate_contrasts(m2, "tx_ori", comparison = "b3 = b2")$Difference,
-    tolerance = 1e-4,
-    ignore_attr = TRUE
-  )
-
-# Hypothesis testing for categorical predictors
-
-A reason to compute adjusted predictions (or estimated marginal means) is to help understanding the relationship between predictors and outcome of a regression model. The next step, which often follows this, is to see if there are statistically significant differences. These could be, for example, differences between groups, i.e. between the levels of categorical predictors or whether trends differ significantly from each other.
-
[…]
