+++ vignettes/introduction_comparisons_2.Rmd
@@ -19,13 +19,8 @@ knitr::opts_chunk$set(
   package.startup.message = FALSE
 )
 options(width = 800)
-arrow_color <- "#FF00cc"
-my_predictions <- NULL
 
 pkgs <- c("marginaleffects", "parameters")
-
-options(modelbased_join_dots = FALSE)
-options(modelbased_select = "minimal")
 
 if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
   knitr::opts_chunk$set(eval = FALSE)
