### 2025-05-12 d1e2ab3
Add vignette on custom contrasts and joint tests (#502)
* Add vignette on custom contrasts and joint tests

* draft

* finalize draft

* fiix

* typos

* styler

* fix

* fix

* add test

* fix

* fix

* news, desc
