+++ vignettes/introduction_comparisons_4.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Contrasts and comparisons for zero-inflation models"
+title: "Contrasts and comparisons for generalized linear models"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Contrasts and comparisons for zero-inflation models}
+  %\VignetteIndexEntry{Contrasts and comparisons for generalized linear models}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -18,8 +18,9 @@ knitr::opts_chunk$set(
 )
 options(width = 800)
 arrow_color <- "#FF00cc"
+p <- ht8 <- NULL
 
-pkgs <- c("ggplot2", "see", "marginaleffects", "glmmTMB")
+pkgs <- c("ggplot2", "see", "marginaleffects")
 
 if (!all(vapply(pkgs, requireNamespace, quietly = TRUE, FUN.VALUE = logical(1L)))) {
   knitr::opts_chunk$set(eval = FALSE)
@@ -27,90 +28,65 @@ if (!all(vapply(pkgs, requireNamespace, quietly = TRUE, FUN.VALUE = logical(1L))
 ```
 
 
-This vignette is the last in a 4-part series:
+This vignette is the fourth in a 5-part series:
 
 1. [**Contrasts and Pairwise Comparisons**](https://easystats.github.io/modelbased/articles/introduction_comparisons_1.html)
 
-2. [**Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**](https://easystats.github.io/modelbased/articles/introduction_comparisons_2.html)
-
-3. [**Contrasts and Comparisons for Generalized Linear Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
-
-4. **Contrasts and Comparisons for Zero-Inflation Models**
+2. [**User Defined Contrasts and Joint Tests**](https://easystats.github.io/modelbased/articles/introduction_comparisons_2.html)
 
 
-# Contrasts and comparisons for Zero-Inflation Models
+3. [**Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
 
-Lastly, we show an example for models with zero-inflation component.
+4. **Contrasts and Comparisons for Generalized Linear Models**
 
-## What is a zero-inflated model?
+5. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_5.html)
 
-A zero-inflated model is a statistical approach used when dealing with count data that has an excessive number of zero values. Imagine counting something that can be zero, like the number of customers a store gets in a day, and it happens that there are a lot more zeros in the data than a typical count model (e.g., Poisson regression) would expect. That's where we need zero-inflated regression models. These models consider two ways zeros can happen:
 
-- True Zeros: These are days the store is naturally closed, or maybe there's just no demand for the product.
+# Contrasts and comparisons for GLM - logistic regression example
 
-- Counting Zeros: These are days the store is open but just happens to get no customers. Maybe it's bad luck, or a random fluctuation.
+We will now show an example for non-Gaussian models. For GLM's (generalized linear models) with (non-Gaussian) link-functions, `estimate_means()` by default returns predicted values on the *response* scale. For example, predicted values for logistic regression models are shown as *probabilities*.
 
[…]
