+++ vignettes/introduction_comparisons_2.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Slopes, floodlight and spotlight analysis (Johnson-Neyman intervals)"
+title: "User Defined Contrasts and Joint Tests"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Slopes, floodlight and spotlight analysis (Johnson-Neyman intervals)}
+  %\VignetteIndexEntry{User Defined Contrasts and Joint Tests}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -19,12 +19,8 @@ knitr::opts_chunk$set(
   package.startup.message = FALSE
 )
 options(width = 800)
-options(modelbased_join_dots = FALSE)
-ht6 <- m <- NULL
 
-arrow_color <- "#FF00cc"
-
-pkgs <- c("ggplot2", "see", "marginaleffects", "parameters")
+pkgs <- c("marginaleffects", "parameters")
 
 if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
   knitr::opts_chunk$set(eval = FALSE)
@@ -32,243 +28,143 @@ if (!all(insight::check_if_installed(pkgs, quietly = TRUE))) {
 ```
 
 
-This vignette is the second in a 4-part series:
+This vignette is the second in a 5-part series:
 
 1. [**Contrasts and Pairwise Comparisons**](https://easystats.github.io/modelbased/articles/introduction_comparisons_1.html)
 
-2. **Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**
+2. **User Defined Contrasts and Joint Tests**
 
-3. [**Contrasts and Comparisons for Generalized Linear Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
+3. [**Comparisons of Slopes, Floodlight and Spotlight Analysis (Johnson-Neyman Intervals)**](https://easystats.github.io/modelbased/articles/introduction_comparisons_3.html)
 
-4. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_4.html)
+4. [**Contrasts and Comparisons for Generalized Linear Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_4.html)
 
-# Contrasts and comparisons for slopes of numeric predictors
+5. [**Contrasts and Comparisons for Zero-Inflation Models**](https://easystats.github.io/modelbased/articles/introduction_comparisons_5.html)
 
-For numeric focal terms, it is possible to calculate contrasts for slopes, or the _linear trend_ of these focal terms. Let's start with a simple example again.
+# User defined contrasts
+
+The first example demonstrates how to define and apply user-defined contrasts to a factor variable within a linear model, respectively how to avoid defining specific contrasts and instead directly formulate the desired contrast as custom hypothesis in `estimate_contrasts()`.
+
+First, we define two specific contrasts: `treat_vs_none` (comparing the average of two treatment levels to a control level) and `short_vs_long` (comparing two treatment levels against each other).
+
+To illustrate the effect of these custom contrasts, the code creates a copy of the original treatment factor (`tx_ori`) before applying the new contrasts to the `tx` factor. It also centers a continuous predictor, `score`, to better interpret the main effects of the interaction term.
+
+Two linear models are then fitted: `m1` uses the `tx` factor with the user-defined contrasts, while `m2` uses the original `tx_ori` factor. The `compare_parameters()` function is used to show that `m1` directly estimates coefficients corresponding to `treat_vs_none` and `short_vs_long`, which are not directly available in `m2`.
+
[…]
