### 2025-05-12 c0fdf99
Deploying to gh-pages from @ easystats/modelbased@d1e2ab3371ffff9e63b77cc248597c0a2d1ad54c 🚀
