### 2025-05-12 #7969 @spholmes
@strengejecke : thanks for making the revisions, I don't think your reference to the iris data is the original one and it would be better to reference Anderson, 1936 (see https://en.wikipedia.org/wiki/Iris_flower_data_set), after you make that reference change, we can move on to the DOI archiving and version stabilization.
