+++ paper/paper.bib
@@ -151,3 +151,18 @@
   doi = {10.1177/25152459241236149},
   URL = {https://doi.org/10.1177/25152459241236149}
 }
+
+@article{anderson_species_1936,
+	title = {The {Species} {Problem} in {Iris}},
+	volume = {23},
+	issn = {00266493},
+	url = {https://www.jstor.org/stable/2394164?origin=crossref},
+	doi = {10.2307/2394164},
+	number = {3},
+	urldate = {2025-05-12},
+	journal = {Annals of the Missouri Botanical Garden},
+	author = {Anderson, Edgar},
+	month = sep,
+	year = {1936},
+	pages = {457},
+}
