### 2025-05-12 ec8cd2d
update md and PDF
