### 2025-05-12 77fdbbc
Deploying to gh-pages from @ easystats/modelbased@ec8cd2df89c4a53b736e4651f3f870637c9d0851 🚀
