### 2025-05-12 #7969 @strengejacke
is osf.io ok, or would you prefer zenodo as archive?
