### 2025-05-12 6c5071f
Set version for JOSS publication
