### 2025-05-12 9d99e9b
Update NEWS.md
