### 2025-05-12 #7969 @strengejacke
- We have updated the version number, the final package/software version for this publication is **0.11.1**
- We have created a GitHub release for this version: https://github.com/easystats/modelbased/releases/tag/v0.11.1
- We created an archive, where the source files for this release can be downloade, the DOI is 10.17605/OSF.IO/XHGZ5 (https://osf.io/xhgz5/)
- The title of the archive matches the title of this publication
- License for archive and software/GitHub is the same
- We checked all ORCID links, all are correct
- We checked all author affiliations, everything is correct
