### 2025-05-12 f506788
Deploying to gh-pages from @ easystats/modelbased@6c5071f1c74a417c97e46c27bc10ae62353ca970 🚀
