### 2025-05-12 05a62ea
Deploying to gh-pages from @ easystats/modelbased@9d99e9bd0f7d36c289ad9c3ce2932d14a11fa99a 🚀
