### 2025-05-12 #7969 @spholmes
> is osf.io ok, or would you prefer zenodo as archive?

Yes, osf.io is OK as long as we have the right DOI. The archive with the relevant DOI should have the same authors and titles, license  as the paper  and the archived package's information from DESCRIPTION/paper.md/LICENSE/etc. You could look at a recent example such as fastfrechet (https://joss.theoj.org/papers/10.21105/joss.07925).
