### 2025-05-12 #7969 @strengejacke
I changed the repository and now use Zenodo, because you can easier add authors with their ORCID than in osf.io. So please use the Zenodo-DOI / URL (I updated the comment)
