### 2025-05-13 f2ed0f1
Minor revision of docs for visualization recipe
