+++ man/visualisation_recipe.estimate_predicted.Rd
@@ -1,11 +1,17 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/visualisation_recipe.R
-\name{visualisation_recipe.estimate_predicted}
+% Please edit documentation in R/plot.R, R/visualisation_recipe.R
+\name{plot.estimate_predicted}
+\alias{plot.estimate_predicted}
+\alias{plot.estimate_means}
 \alias{visualisation_recipe.estimate_predicted}
 \alias{visualisation_recipe.estimate_slopes}
 \alias{visualisation_recipe.estimate_grouplevel}
 \title{Automated plotting for 'modelbased' objects}
 \usage{
+\method{plot}{estimate_predicted}(x, ...)
+
+\method{plot}{estimate_means}(x, ...)
+
 \method{visualisation_recipe}{estimate_predicted}(
   x,
   show_data = FALSE,
@@ -43,8 +49,11 @@
 \arguments{
 \item{x}{A modelbased object.}
 
+\item{...}{Arguments passed from \code{plot()} to \code{visualisation_recipe()}.}
+
 \item{show_data}{Logical, if \code{TRUE}, display the "raw" data as a background
-to the model-based estimation.}
+to the model-based estimation. This argumenmt will be ignored for plotting
+objects returned by \code{estimate_slopes()} or \code{estimate_grouplevel()}.}
 
 \item{point, line, pointrange, ribbon, facet, grid}{Additional
 aesthetics and parameters for the geoms (see customization example).}
@@ -64,25 +73,30 @@ deviation around the mean are chosen as representative values for that
 predictor. Use \code{FALSE} to always use continuous color scales for numeric
 predictors. It is possible to set a global default value using \code{options()},
 e.g. \code{options(modelbased_numeric_as_discrete = 10)}.}
-
-\item{...}{Not used.}
+}
+\value{
+An object of class \code{visualisation_recipe} that describes the layers
+used to create a plot based on \verb{\{ggplot2\}. The related }plot()\verb{method is in the}{see}` package.
 }
 \description{
-Most 'modelbased' objects can be visualized using the \code{plot()} function, which
-internally calls the \code{visualisation_recipe()} function. See the \strong{examples}
-below for more information and examples on how to create and customize plots.
-}
-\details{
-The plotting works by mapping any predictors from the \code{by} argument to the x-axis,
-colors, alpha (transparency) and facets. Thus, the appearance of the plot depends
-on the order of the variables that you specify in the \code{by} argument. For instance,
-the plots corresponding to \code{estimate_relation(model, by=c("Species", "Sepal.Length"))}
-and \code{estimate_relation(model, by=c("Sepal.Length", "Species"))} will look different.
+Most \strong{modelbased} objects can be visualized using the \code{plot()} function,
+which internally calls the \code{visualisation_recipe()} function. See the
+examples below for more information and examples on how to create and
+customize plots.
+
+The plotting works by mapping any predictors from the \code{by} argument to the
+x-axis, colors, alpha (transparency) and facets. Thus, the appearance of the
+plot depends on the order of the variables that you specify in the \code{by}
+argument. For instance, the plots corresponding to
+\code{estimate_relation(model, by=c("Species", "Sepal.Length"))} and
+\code{estimate_relation(model, by=c("Sepal.Length", "Species"))} will look
+different.
 
 The automated plotting is primarily meant for convenient visual checks, but
 for publication-ready figures, we recommend re-creating the figures using the
-\code{ggplot2} package directly.
-
+\code{{ggplot2}} package directly.
+}
+\details{
 There are two options to remove the confidence bands or errors bars
 from the plot. To remove error bars, simply set the \code{pointrange} geom to
 \code{point}, e.g. \code{plot(..., pointrange = list(geom = "point"))}. To remove the
