### 2025-05-13 #7969 @spholmes
@strengejacke : We're almost there, could you add the DOI for the Pan and Huang paper?
Thanks
