### 2025-05-13 d8c5477
Add doi for Pan, Huang
