+++ paper/paper.bib
@@ -16,6 +16,7 @@
   volume={24},
   pages={725--738},
   year={2014},
+  doi = {10.1007/s11222-013-9398-0},
   publisher={Springer}
 }
 
