### 2025-05-13 #7969 @strengejacke
Sure, I updated the `.bib` file and checked the PDF locally, DOI now added to the reference.
