### 2025-05-13 #7969 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
