### 2025-05-13 65ebbc0
`estimate_contrasts()` for `estimate_relations()` works for four focal terms and is more efficient
