+++ tests/testthat/_snaps/estimate_contrasts_methods.md
@@ -86,26 +86,26 @@
     Output
       Model-based Contrasts Analysis
       
-      age   | Parameter             | Difference |   SE |         95% CI | Statistic |     p
-      --------------------------------------------------------------------------------------
-      -40   | Male no -Female no    |       0.07 | 1.11 | [-2.11,  2.25] |      0.06 | 0.951
-      -40   | Male no -Male yes     |      -0.88 | 1.18 | [-3.18,  1.43] |     -0.75 | 0.455
-      -40   | Female no -Male yes   |      -0.95 | 1.04 | [-2.99,  1.10] |     -0.91 | 0.364
-      -40   | Male no -Female yes   |      -0.79 | 1.07 | [-2.89,  1.30] |     -0.74 | 0.457
-      -40   | Female no -Female yes |      -0.86 | 0.92 | [-2.67,  0.94] |     -0.94 | 0.348
-      -40   | Male yes -Female yes  |       0.08 | 0.99 | [-1.86,  2.03] |      0.08 | 0.932
-      41-64 | Male no -Female no    |       1.18 | 0.74 | [-0.26,  2.63] |      1.60 | 0.109
-      41-64 | Male no -Male yes     |      -0.53 | 0.86 | [-2.22,  1.16] |     -0.62 | 0.538
-      41-64 | Female no -Male yes   |      -1.71 | 0.65 | [-2.98, -0.44] |     -2.65 | 0.008
-      41-64 | Male no -Female yes   |       0.84 | 0.74 | [-0.62,  2.30] |      1.13 | 0.258
-      41-64 | Female no -Female yes |      -0.34 | 0.48 | [-1.28,  0.60] |     -0.71 | 0.476
-      41-64 | Male yes -Female yes  |       1.37 | 0.66 | [ 0.09,  2.66] |      2.09 | 0.036
-      65+   | Male no -Female no    |       0.88 | 0.77 | [-0.62,  2.38] |      1.15 | 0.250
-      65+   | Male no -Male yes     |      -0.93 | 1.25 | [-3.39,  1.53] |     -0.74 | 0.459
-      65+   | Female no -Male yes   |      -1.81 | 1.17 | [-4.10,  0.48] |     -1.55 | 0.122
-      65+   | Male no -Female yes   |      -0.44 | 1.21 | [-2.82,  1.94] |     -0.36 | 0.717
-      65+   | Female no -Female yes |      -1.32 | 1.13 | [-3.53,  0.89] |     -1.17 | 0.241
-      65+   | Male yes -Female yes  |       0.49 | 1.50 | [-2.45,  3.43] |      0.33 | 0.745
+      Level1     | Level2      | age   | Difference |   SE |         95% CI | Statistic |     p
+      -----------------------------------------------------------------------------------------
+      Male, no   | Female, no  | -40   |       0.07 | 1.11 | [-2.11,  2.25] |      0.06 | 0.951
+      Male, no   | Male, yes   | -40   |      -0.88 | 1.18 | [-3.18,  1.43] |     -0.75 | 0.455
+      Female, no | Male, yes   | -40   |      -0.95 | 1.04 | [-2.99,  1.10] |     -0.91 | 0.364
+      Male, no   | Female, yes | -40   |      -0.79 | 1.07 | [-2.89,  1.30] |     -0.74 | 0.457
+      Female, no | Female, yes | -40   |      -0.86 | 0.92 | [-2.67,  0.94] |     -0.94 | 0.348
+      Male, yes  | Female, yes | -40   |       0.08 | 0.99 | [-1.86,  2.03] |      0.08 | 0.932
[…]
