+++ R/estimate_contrast_methods.R
@@ -33,12 +33,17 @@ estimate_contrasts.estimate_predicted <- function(model,
     ))
   }
 
-  # sanity check
+  # sanity check - if user did not provide contrasts, we assume that all
+  # focal terms are contrasts
   if (is.null(contrast)) {
     contrast <- focal_terms
     if (!is.null(by)) {
       contrast <- setdiff(contrast, by)
     }
+  } else if (is.null(by)) {
+    # else, if contrasts are specified, but not `by`, we set by to the
+    # remaining focal terms
+    by <- setdiff(focal_terms, contrast)
   }
 
   # sanity check - user-defined by-variables may not be in the data
