+++ R/estimate_means.R
@@ -402,11 +402,7 @@ estimate_means <- function(model,
   }
 
   # sanity check - did method return standard errors?
-  if (verbose && !is.null(means$SE) && isTRUE(all(means$SE == means$SE[1]))) {
-    insight::format_alert(
-      "Standard errors are probably not reliable. This can happen when random effects are involved. You may try `estimate_relation()` instead." # nolint
-    )
-  }
+  .check_standard_errors(out = means, model = model, verbose = verbose)
 
   # restore attributes later
   info <- attributes(estimated)
