### 2025-05-20 459c176
`estimate_contrasts()` for `estimate_relations()` works for four focal terms and is more efficient (#504)
* `estimate_contrasts()` for `estimate_relations()` works for four focal terms and is more efficient

* more flexible

* add test

* fix

* fix snapshot

* add tests

* add test

* news, desc

* msg

* msg, test

* typo2

* fix, add tests

* comment code

* fix, add test

* update examples

* warning

* update

* Update format.R

* Update format.R

* Update format.R

* Update estimate_contrasts.md

* allow to pass tolerance, increase default tolerance
