### 2025-05-20 f32479d
use puppy_love data, cite it
