+++ pkgdown/_pkgdown.yml
@@ -29,7 +29,7 @@ reference:
   - title: "Sample datasets"
     contents:
       - coffee_data
-      - contrast_example
+      - puppy_love
       - efc
       - fish
 
