### 2025-05-20 a612ef0
Deploying to gh-pages from @ easystats/modelbased@f32479dddf5f934918c12487e8e209deafcb744d 🚀
