+++ pkgdown.yml
@@ -1,6 +1,6 @@
-pandoc: 3.6.4
-pkgdown: 2.1.2.9000
-pkgdown_sha: 99457e7fa7f36696ae743b06aca59c0fdd97a7d6
+pandoc: 3.7.0.1
+pkgdown: 2.1.3.9000
+pkgdown_sha: f950aed41af49c732ba7884d0806c3641a7b94b3
 articles:
   derivatives: derivatives.html
   estimate_contrasts: estimate_contrasts.html
@@ -22,7 +22,7 @@ articles:
   practical_intersectionality: practical_intersectionality.html
   visualisation_matrix: visualisation_matrix.html
   workflow_modelbased: workflow_modelbased.html
-last_built: 2025-05-13T07:02Z
+last_built: 2025-05-20T14:11Z
 urls:
   reference: https://easystats.github.io/modelbased/reference
   article: https://easystats.github.io/modelbased/articles
