### 2025-05-20 48c5b8f
Deploying to gh-pages from @ easystats/modelbased@3e9f55a26f3121b0a967e49e0bea38a9f40ae0ff 🚀
