### 2025-05-20 dc3d08b
Deploying to gh-pages from @ easystats/modelbased@9814742ba6bd2536009ad77fb50cb3e94cd67c3f 🚀
