+++ DESCRIPTION
@@ -39,7 +39,7 @@ BugReports: https://github.com/easystats/modelbased/issues
 Depends:
     R (>= 3.6)
 Imports:
-    bayestestR (>= 0.15.3),
+    bayestestR (>= 0.16.0),
     datawizard (>= 1.0.2),
     insight (>= 1.3.0),
     parameters (>= 0.25.0),
