### 2025-05-20 2cd31bc
Add tests for survival models (#506)
* Add tests for survival

* desc

* bayestestR
