+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: modelbased
 Title: Estimation of Model-Based Predictions, Contrasts and Means
-Version: 0.11.1.1
+Version: 0.11.1.2
 Authors@R:
     c(person(given = "Dominique",
              family = "Makowski",
@@ -39,7 +39,7 @@ BugReports: https://github.com/easystats/modelbased/issues
 Depends:
     R (>= 3.6)
 Imports:
-    bayestestR (>= 0.15.3),
+    bayestestR (>= 0.16.0),
     datawizard (>= 1.0.2),
     insight (>= 1.3.0),
     parameters (>= 0.25.0),
@@ -91,6 +91,7 @@ Suggests:
     RWiener,
     sandwich,
     see (>= 0.11.0),
+    survival,
     testthat (>= 3.2.1),
     vdiffr,
     withr
