### 2025-05-21 #7969 @DominiqueMakowski
In case it helps, we double checked the proof and all seems error-free :)
