### 2025-05-22 f36e6e1
Deploying to gh-pages from @ easystats/modelbased@681de9281bd04f3023de9604de8250eedb3a0c2b 🚀
